+++ support/docker/postgres.go
@@ -125,6 +125,10 @@ func (r *PostgresImpl) Ready() error {
 	return r.close(gormDB)
 }
 
+func (r *PostgresImpl) Reuse(containerID string, port int) error {
+	return nil
+}
+
 func (r *PostgresImpl) Shutdown() error {
 	if _, err := run(fmt.Sprintf("docker stop %s", r.containerID)); err != nil {
 		return fmt.Errorf("stop Postgres error: %v", err)
