+++ tests/orm_test.go
@@ -1,3888 +1,168 @@
 package tests
 
 import (
-	"fmt"
-	"strconv"
+	"context"
+	"errors"
 	"testing"
-	"time"
+
+	"github.com/stretchr/testify/suite"
 
 	contractsorm "github.com/goravel/framework/contracts/database/orm"
 	"github.com/goravel/framework/database/gorm"
-	"github.com/goravel/framework/errors"
-	mocksconfig "github.com/goravel/framework/mocks/config"
-	"github.com/goravel/framework/support/carbon"
-	"github.com/goravel/framework/testing/utils"
-	"github.com/goravel/postgres"
-	"github.com/stretchr/testify/suite"
+	"github.com/goravel/framework/database/orm"
 )
 
-const (
-	DriverPostgres  = "postgres"
-	DriverMysql     = "mysql"
-	DriverSqlite    = "sqlite"
-	DriverSqlserver = "sqlserver"
-)
-
-type OrmTestSuite struct {
+type OrmSuite struct {
 	suite.Suite
-	database string
-	username string
-	password string
-	queries  map[string]*gorm.TestQuery1
+	orm           *orm.Orm
+	defaultDriver string
+	queries       map[string]*gorm.TestQuery1
 }
 
-func TestOrmTestSuite(t *testing.T) {
-	suite.Run(t, &OrmTestSuite{
-		database: "postgres",
-		username: "postgres",
-		password: "Framework!123",
-		queries:  make(map[string]*gorm.TestQuery1),
+func TestOrmSuite(t *testing.T) {
+	suite.Run(t, &OrmSuite{
+		queries: make(map[string]*gorm.TestQuery1),
 	})
 }
 
-func (s *OrmTestSuite) SetupSuite() {
-	postgresTestQuery := s.postgresTestQuery()
-	s.queries[postgresTestQuery.Driver().Config().Driver] = postgresTestQuery
+func (s *OrmSuite) SetupSuite() {
+	postgresTestQuery := postgresTestQuery("", false)
+	s.queries[postgresTestQuery.Driver().Config().Connection] = postgresTestQuery
+	s.defaultDriver = postgresTestQuery.Driver().Config().Driver
 }
 
-func (s *OrmTestSuite) SetupTest() {
+func (s *OrmSuite) SetupTest() {
+	queries := make(map[string]contractsorm.Query)
+
+	for driver, query := range s.queries {
+		queries[driver] = query.Query()
+	}
+
+	s.orm = orm.NewOrm(context.Background(), nil, s.defaultDriver, queries[s.defaultDriver], queries, nil, nil, nil)
 }
 
-func (s *OrmTestSuite) TearDownSuite() {
+func (s *OrmSuite) TearDownSuite() {
 	// TODO Shutdown Sqlite
 	// if s.queries[database.DriverSqlite] != nil {
 	// 	s.NoError(s.queries[database.DriverSqlite].Docker().Shutdown())
 	// }
 }
 
-func (s *OrmTestSuite) postgresTestQuery() *gorm.TestQuery1 {
-	docker := postgres.NewDocker(s.database, s.username, s.password)
-	if err := docker.Build(); err != nil {
-		panic(err)
+func (s *OrmSuite) TestConnection() {
+	for connection := range s.queries {
+		s.NotNil(s.orm.Connection(connection))
 	}
-	if err := docker.Ready(); err != nil {
-		panic(err)
-	}
-
-	mockConfig := &mocksconfig.Config{}
-
-	mockConfig.EXPECT().GetBool("app.debug").Return(true)
-	mockConfig.EXPECT().GetInt("database.slow_threshold", 200).Return(200)
-	mockConfig.EXPECT().GetInt("database.pool.max_idle_conns", 10).Return(10)
-	mockConfig.EXPECT().GetInt("database.pool.max_open_conns", 100).Return(100)
-	mockConfig.EXPECT().GetInt("database.pool.conn_max_idletime", 3600).Return(3600)
-	mockConfig.EXPECT().GetInt("database.pool.conn_max_lifetime", 3600).Return(3600)
-
-	mockConfig.EXPECT().Get("database.connections.postgres.read").Return(nil)
-	mockConfig.EXPECT().Get("database.connections.postgres.write").Return(nil)
-	mockConfig.EXPECT().GetString("database.connections.postgres.host").Return("localhost")
-	mockConfig.EXPECT().GetInt("database.connections.postgres.port").Return(docker.Config().Port)
-	mockConfig.EXPECT().GetString("database.connections.postgres.username").Return(s.username)
-	mockConfig.EXPECT().GetString("database.connections.postgres.password").Return(s.password)
-	mockConfig.EXPECT().GetString("database.connections.postgres.database").Return(s.database)
-	mockConfig.EXPECT().GetString("database.connections.postgres.sslmode").Return("disable")
-	mockConfig.EXPECT().GetString("database.connections.postgres.timezone").Return("UTC")
-	mockConfig.EXPECT().GetString("database.connections.postgres.prefix").Return("")
-	mockConfig.EXPECT().GetBool("database.connections.postgres.singular").Return(false)
-	mockConfig.EXPECT().GetBool("database.connections.postgres.no_lower_case").Return(false)
-	mockConfig.EXPECT().GetString("database.connections.postgres.dsn").Return("")
-	mockConfig.EXPECT().GetString("database.connections.postgres.schema", "public").Return("public")
-	mockConfig.EXPECT().Get("database.connections.postgres.name_replacer").Return(nil)
-
-	driver := postgres.NewPostgres(mockConfig, utils.NewTestLog(), "postgres")
-
-	testQuery, err := gorm.NewTestQuery1(driver, mockConfig)
-	if err != nil {
-		panic(err)
-	}
-
-	testQuery.CreateTable()
-
-	return testQuery
 }
 
-func (s *OrmTestSuite) TestAssociation() {
-	for driver, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "Find",
-				setup: func() {
-					user := User{
-						Name: "association_find_name",
-						Address: &Address{
-							Name: "association_find_address",
-						},
-						age: 1,
-					}
+func (s *OrmSuite) TestDB() {
+	db, err := s.orm.DB()
+	s.NotNil(db)
+	s.Nil(err)
 
-					s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID > 0)
+	for connection := range s.queries {
+		db, err := s.orm.Connection(connection).DB()
+		s.NotNil(db)
+		s.Nil(err)
+	}
+}
 
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
+func (s *OrmSuite) TestQuery() {
+	s.NotNil(s.orm.Query())
 
-					var userAddress Address
-					s.Nil(query.Query().Model(&user1).Association("Address").Find(&userAddress))
-					s.True(userAddress.ID > 0)
-					s.Equal("association_find_address", userAddress.Name)
-				},
-			},
-			{
-				name: "hasOne Append",
-				setup: func() {
-					user := User{
-						Name: "association_has_one_append_name",
-						Address: &Address{
-							Name: "association_has_one_append_address",
-						},
-					}
-
-					s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID), driver)
-					s.True(user1.ID > 0, driver)
-					s.Nil(query.Query().Model(&user1).Association("Address").Append(&Address{Name: "association_has_one_append_address1"}), driver)
-
-					s.Nil(query.Query().Load(&user1, "Address"), driver)
-					s.True(user1.Address.ID > 0, driver)
-					s.Equal("association_has_one_append_address1", user1.Address.Name, driver)
-				},
-			},
-			{
-				name: "hasMany Append",
-				setup: func() {
-					user := User{
-						Name: "association_has_many_append_name",
-						Books: []*Book{
-							{Name: "association_has_many_append_address1"},
-							{Name: "association_has_many_append_address2"},
-						},
-					}
-
-					s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Books[0].ID > 0)
-					s.True(user.Books[1].ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Nil(query.Query().Model(&user1).Association("Books").Append(&Book{Name: "association_has_many_append_address3"}))
-
-					s.Nil(query.Query().Load(&user1, "Books"))
-					s.Equal(3, len(user1.Books))
-					s.Equal("association_has_many_append_address3", user1.Books[2].Name)
-				},
-			},
-			{
-				name: "hasOne Replace",
-				setup: func() {
-					user := User{
-						Name: "association_has_one_append_name",
-						Address: &Address{
-							Name: "association_has_one_append_address",
-						},
-					}
-
-					s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Nil(query.Query().Model(&user1).Association("Address").Replace(&Address{Name: "association_has_one_append_address1"}))
-
-					s.Nil(query.Query().Load(&user1, "Address"))
-					s.True(user1.Address.ID > 0)
-					s.Equal("association_has_one_append_address1", user1.Address.Name)
-				},
-			},
-			{
-				name: "hasMany Replace",
-				setup: func() {
-					user := User{
-						Name: "association_has_many_replace_name",
-						Books: []*Book{
-							{Name: "association_has_many_replace_address1"},
-							{Name: "association_has_many_replace_address2"},
-						},
-					}
-
-					s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Books[0].ID > 0)
-					s.True(user.Books[1].ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Nil(query.Query().Model(&user1).Association("Books").Replace(&Book{Name: "association_has_many_replace_address3"}))
-
-					s.Nil(query.Query().Load(&user1, "Books"))
-					s.Equal(1, len(user1.Books))
-					s.Equal("association_has_many_replace_address3", user1.Books[0].Name)
-				},
-			},
-			{
-				name: "Delete",
-				setup: func() {
-					user := User{
-						Name: "association_delete_name",
-						Address: &Address{
-							Name: "association_delete_address",
-						},
-					}
-
-					s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID > 0)
-
-					// No ID when Delete
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Nil(query.Query().Model(&user1).Association("Address").Delete(&Address{Name: "association_delete_address"}))
-
-					s.Nil(query.Query().Load(&user1, "Address"))
-					s.True(user1.Address.ID > 0)
-					s.Equal("association_delete_address", user1.Address.Name)
-
-					// Has ID when Delete
-					var user2 User
-					s.Nil(query.Query().Find(&user2, user.ID))
-					s.True(user2.ID > 0)
-					var userAddress Address
-					userAddress.ID = user1.Address.ID
-					s.Nil(query.Query().Model(&user2).Association("Address").Delete(&userAddress))
-
-					s.Nil(query.Query().Load(&user2, "Address"))
-					s.Nil(user2.Address)
-				},
-			},
-			{
-				name: "Clear",
-				setup: func() {
-					user := User{
-						Name: "association_clear_name",
-						Address: &Address{
-							Name: "association_clear_address",
-						},
-					}
-
-					s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID > 0)
-
-					// No ID when Delete
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Nil(query.Query().Model(&user1).Association("Address").Clear())
-
-					s.Nil(query.Query().Load(&user1, "Address"))
-					s.Nil(user1.Address)
-				},
-			},
-			{
-				name: "Count",
-				setup: func() {
-					user := User{
-						Name: "association_count_name",
-						Books: []*Book{
-							{Name: "association_count_address1"},
-							{Name: "association_count_address2"},
-						},
-					}
-
-					s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Books[0].ID > 0)
-					s.True(user.Books[1].ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Equal(int64(2), query.Query().Model(&user1).Association("Books").Count())
-				},
-			},
+	s.NotPanics(func() {
+		for i := 0; i < 5; i++ {
+			go func() {
+				var role Role
+				_ = s.orm.Query().Find(&role, 1)
+			}()
 		}
+	})
 
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
+	for connection := range s.queries {
+		s.NotNil(s.orm.Connection(connection).Query())
 	}
 }
 
-func (s *OrmTestSuite) TestBelongsTo() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := &User{
-				Name: "belongs_to_name",
-				Address: &Address{
-					Name: "belongs_to_address",
-				},
-			}
+func (s *OrmSuite) TestFactory() {
+	s.NotNil(s.orm.Factory())
 
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-			s.True(user.Address.ID > 0)
-
-			var userAddress Address
-			s.Nil(query.Query().With("User").Where("name = ?", "belongs_to_address").First(&userAddress))
-			s.True(userAddress.ID > 0)
-			s.True(userAddress.User.ID > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestCount() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "count_user", Avatar: "count_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "count_user", Avatar: "count_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var count int64
-			s.Nil(query.Query().Model(&User{}).Where("name = ?", "count_user").Count(&count))
-			s.True(count > 0)
-
-			var count1 int64
-			s.Nil(query.Query().Table("users").Where("name = ?", "count_user").Count(&count1))
-			s.True(count1 > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestCreate() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "success by struct",
-				setup: func() {
-					user := User{Name: "create_user"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-				},
-			},
-			{
-				name: "batch create success by struct",
-				setup: func() {
-					users := []User{
-						{Name: "batch_create_user_by_struct_1"},
-						{Name: "batch_create_user_by_struct_2"},
-					}
-					s.Nil(query.Query().Create(&users))
-					s.True(users[0].ID > 0)
-					s.True(users[1].ID > 0)
-				},
-			},
-			{
-				name: "success by map",
-				setup: func() {
-					s.Nil(query.Query().Table("users").Create(map[string]any{
-						"name":       "create_by_map_name1",
-						"avatar":     "create_by_map_avatar1",
-						"created_at": carbon.Now(),
-						"updated_at": carbon.Now(),
-					}))
-
-					var user1 User
-					err := query.Query().Where("name", "create_by_map_name1").
-						Where("avatar", "create_by_map_avatar1").First(&user1)
-					s.NoError(err)
-					s.True(user1.ID > 0)
-
-					s.Nil(query.Query().Model(User{}).Create(map[string]any{
-						"Name":      "create_by_map_name2",
-						"Avatar":    "create_by_map_avatar2",
-						"CreatedAt": carbon.Now(),
-						"UpdatedAt": carbon.Now(),
-					}))
-
-					var user2 User
-					err = query.Query().Where("name", "create_by_map_name2").
-						Where("avatar", "create_by_map_avatar2").First(&user2)
-					s.NoError(err)
-					s.True(user2.ID > 0)
-				},
-			},
-			{
-				name: "batch create success by map",
-				setup: func() {
-					s.Nil(query.Query().Table("users").Create([]map[string]any{
-						{
-							"name":       "batch_create_by_map_name1",
-							"avatar":     "batch_create_by_map_avatar1",
-							"created_at": carbon.Now(),
-							"updated_at": carbon.Now(),
-						},
-						{
-							"name":       "batch_create_by_map_name2",
-							"avatar":     "batch_create_by_map_avatar2",
-							"created_at": carbon.Now(),
-							"updated_at": carbon.Now(),
-						},
-					}))
-
-					var users1 []User
-					err := query.Query().Where("name", "batch_create_by_map_name1").OrWhere("name", "batch_create_by_map_name2").Find(&users1)
-					s.NoError(err)
-					s.Len(users1, 2)
-
-					// The []map should be a pointer, otherwise gorm will throw an error
-					s.Nil(query.Query().Model(User{}).Create(&[]map[string]any{
-						{
-							"Name":      "batch_create_by_map_name3",
-							"Avatar":    "batch_create_by_map_avatar3",
-							"CreatedAt": carbon.Now(),
-							"UpdatedAt": carbon.Now(),
-						},
-						{
-							"Name":      "batch_create_by_map_name4",
-							"Avatar":    "batch_create_by_map_avatar4",
-							"CreatedAt": carbon.Now(),
-							"UpdatedAt": carbon.Now(),
-						},
-					}))
-
-					var users2 []User
-					err = query.Query().Where("name", "batch_create_by_map_name3").OrWhere("name", "batch_create_by_map_name4").Find(&users2)
-					s.NoError(err)
-					s.Len(users2, 2)
-				},
-			},
-			// {
-			// 	name: "success when refresh connection",
-			// 	setup: func() {
-			// 		mockCommonConnection(query.MockConfig(), s.additionalQuery, "dummy")
-
-			// 		people := People{Body: "create_people"}
-			// 		s.Nil(query.Query().Create(&people))
-			// 		s.True(people.ID > 0)
-
-			// 		var count int64
-			// 		err := query.Query().Table("peoples").Where("body", "create_people").Count(&count)
-			// 		s.NoError(err)
-			// 		s.True(count == 0)
-
-			// 		s.Nil(query.Query().Model(&People{}).Create(map[string]any{
-			// 			"body":       "create_people1",
-			// 			"created_at": carbon.Now(),
-			// 			"updated_at": carbon.Now(),
-			// 		}))
-
-			// 		var people1 People
-			// 		s.Nil(query.Query().Where("body", "create_people1").First(&people1))
-			// 		s.True(people1.ID > 0)
-
-			// 		err = query.Query().Table("peoples").Where("body", "create_people1").Count(&count)
-			// 		s.NoError(err)
-			// 		s.True(count == 0)
-			// 	},
-			// },
-			{
-				name: "success when create with no relationships",
-				setup: func() {
-					user := User{Name: "create_user", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "create_address"
-					user.Books[0].Name = "create_book0"
-					user.Books[1].Name = "create_book1"
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID == 0)
-					s.True(user.Books[0].ID == 0)
-					s.True(user.Books[1].ID == 0)
-				},
-			},
-			{
-				name: "success when create with select Associations",
-				setup: func() {
-					user := User{Name: "create_user", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "create_address"
-					user.Books[0].Name = "create_book0"
-					user.Books[1].Name = "create_book1"
-					s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID > 0)
-					s.True(user.Books[0].ID > 0)
-					s.True(user.Books[1].ID > 0)
-				},
-			},
-			{
-				name: "success when create with select fields",
-				setup: func() {
-					user := User{Name: "create_user", Avatar: "create_avatar", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "create_address"
-					user.Books[0].Name = "create_book0"
-					user.Books[1].Name = "create_book1"
-					s.Nil(query.Query().Select("Name", "Avatar", "Address").Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID > 0)
-					s.True(user.Books[0].ID == 0)
-					s.True(user.Books[1].ID == 0)
-				},
-			},
-			{
-				name: "success when create with omit fields",
-				setup: func() {
-					user := User{Name: "create_user", Avatar: "create_avatar", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "create_address"
-					user.Books[0].Name = "create_book0"
-					user.Books[1].Name = "create_book1"
-					s.Nil(query.Query().Omit("Address").Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID == 0)
-					s.True(user.Books[0].ID > 0)
-					s.True(user.Books[1].ID > 0)
-				},
-			},
-			{
-				name: "success create with omit Associations",
-				setup: func() {
-					user := User{Name: "create_user", Avatar: "create_avatar", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "create_address"
-					user.Books[0].Name = "create_book0"
-					user.Books[1].Name = "create_book1"
-					s.Nil(query.Query().Omit(gorm.Associations).Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID == 0)
-					s.True(user.Books[0].ID == 0)
-					s.True(user.Books[1].ID == 0)
-				},
-			},
-			{
-				name: "error when set select and omit at the same time",
-				setup: func() {
-					user := User{Name: "create_user", Avatar: "create_avatar", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "create_address"
-					user.Books[0].Name = "create_book0"
-					user.Books[1].Name = "create_book1"
-					s.EqualError(query.Query().Omit(gorm.Associations).Select("Name").Create(&user), errors.OrmQuerySelectAndOmitsConflict.Error())
-				},
-			},
-			{
-				name: "error when select that set fields and Associations at the same time",
-				setup: func() {
-					user := User{Name: "create_user", Avatar: "create_avatar", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "create_address"
-					user.Books[0].Name = "create_book0"
-					user.Books[1].Name = "create_book1"
-					s.EqualError(query.Query().Select("Name", gorm.Associations).Create(&user), errors.OrmQueryAssociationsConflict.Error())
-				},
-			},
-			{
-				name: "error when omit that set fields and Associations at the same time",
-				setup: func() {
-					user := User{Name: "create_user", Avatar: "create_avatar", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "create_address"
-					user.Books[0].Name = "create_book0"
-					user.Books[1].Name = "create_book1"
-					s.EqualError(query.Query().Omit("Name", gorm.Associations).Create(&user), errors.OrmQueryAssociationsConflict.Error())
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestCursor() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "cursor_user", Avatar: "cursor_avatar", Address: &Address{Name: "cursor_address"}, Books: []*Book{
-				{Name: "cursor_book"},
-			}}
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "cursor_user", Avatar: "cursor_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			user2 := User{Name: "cursor_user", Avatar: "cursor_avatar2"}
-			s.Nil(query.Query().Create(&user2))
-			s.True(user2.ID > 0)
-			res, err := query.Query().Delete(&user2)
-			s.Nil(err)
-			s.Equal(int64(1), res.RowsAffected)
-
-			users, err := query.Query().Model(&User{}).Where("name = ?", "cursor_user").WithTrashed().With("Address").With("Books").Cursor()
-			s.Nil(err)
-			var size int
-			var addressNum int
-			var bookNum int
-			for row := range users {
-				var tempUser User
-				s.Nil(row.Scan(&tempUser))
-				s.True(tempUser.ID > 0)
-				s.True(len(tempUser.Name) > 0)
-				s.NotEmpty(tempUser.CreatedAt.String())
-				s.NotEmpty(tempUser.UpdatedAt.String())
-				s.Equal(tempUser.DeletedAt.Valid, tempUser.ID == user2.ID)
-				size++
-
-				if tempUser.Address != nil {
-					addressNum++
-				}
-				bookNum += len(tempUser.Books)
-			}
-			s.Equal(3, size)
-			s.Equal(1, addressNum)
-			s.Equal(1, bookNum)
-		})
-	}
-}
-
-// func (s *OrmTestSuite) TestDBRaw() {
-// 	userName := "db_raw"
-// 	for driver, query := range s.queries {
-// 		s.Run(driver, func() {
-// 			user := User{Name: userName}
-
-// 			s.Nil(query.Query().Create(&user))
-// 			s.True(user.ID > 0)
-// 			switch driver {
-// 			case database.DriverSqlserver, database.DriverMysql:
-// 				res, err := query.Query().Model(&user).Update("Name", databasedb.Raw("concat(name, ?)", driver))
-// 				s.Nil(err)
-// 				s.Equal(int64(1), res.RowsAffected)
-// 			default:
-// 				res, err := query.Query().Model(&user).Update("Name", databasedb.Raw("name || ?", driver))
-// 				s.Nil(err)
-// 				s.Equal(int64(1), res.RowsAffected)
-// 			}
-
-// 			var user1 User
-// 			s.Nil(query.Query().Find(&user1, user.ID))
-// 			s.True(user1.ID > 0)
-// 			s.True(user1.Name == userName+driver)
-// 		})
-// 	}
-// }
-
-func (s *OrmTestSuite) TestDelete() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "success",
-				setup: func() {
-					user := User{Name: "delete_user", Avatar: "delete_avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					res, err := query.Query().Delete(&user)
-					s.Equal(int64(1), res.RowsAffected)
-					s.Nil(err)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal(uint(0), user1.ID)
-				},
-			},
-			{
-				name: "success by table",
-				setup: func() {
-					user := User{Name: "delete_user_by_table", Avatar: "delete_avatar_by_table"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					res, err := query.Query().Table("users").Where("name", "delete_user_by_table").Delete()
-					s.Equal(int64(1), res.RowsAffected)
-					s.Nil(err)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal(uint(0), user1.ID)
-				},
-			},
-			{
-				name: "success by model",
-				setup: func() {
-					user := User{Name: "delete_user_by_model", Avatar: "delete_avatar_by_model"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					res, err := query.Query().Model(&User{}).Where("name", "delete_user_by_model").Delete()
-					s.Equal(int64(1), res.RowsAffected)
-					s.Nil(err)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal(uint(0), user1.ID)
-				},
-			},
-			// {
-			// 	name: "success when refresh connection",
-			// 	setup: func() {
-			// 		user := User{Name: "delete_user", Avatar: "delete_avatar"}
-			// 		s.Nil(query.Query().Create(&user))
-			// 		s.True(user.ID > 0)
-
-			// 		res, err := query.Query().Delete(&user)
-			// 		s.Equal(int64(1), res.RowsAffected)
-			// 		s.Nil(err)
-
-			// 		var user1 User
-			// 		s.Nil(query.Query().Find(&user1, user.ID))
-			// 		s.Equal(uint(0), user1.ID)
-
-			// 		// refresh connection
-			// 		mockCommonConnection(query.MockConfig(), query, "dummy")
-
-			// 		people := People{Body: "delete_people"}
-			// 		s.Nil(query.Query().Create(&people))
-			// 		s.True(people.ID > 0)
-
-			// 		res, err = query.Query().Delete(&people)
-			// 		s.Equal(int64(1), res.RowsAffected)
-			// 		s.Nil(err)
-
-			// 		var people1 People
-			// 		s.Nil(query.Query().Find(&people1, people.ID))
-			// 		s.Equal(uint(0), people1.ID)
-			// 	},
-			// },
-			{
-				name: "success by id",
-				setup: func() {
-					user := User{Name: "delete_user", Avatar: "delete_avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					res, err := query.Query().Where("id", user.ID).Delete(&User{})
-					s.Equal(int64(1), res.RowsAffected)
-					s.Nil(err)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal(uint(0), user1.ID)
-				},
-			},
-			{
-				name: "success by multiple",
-				setup: func() {
-					users := []User{{Name: "delete_user", Avatar: "delete_avatar"}, {Name: "delete_user1", Avatar: "delete_avatar1"}}
-					s.Nil(query.Query().Create(&users))
-					s.True(users[0].ID > 0)
-					s.True(users[1].ID > 0)
-
-					res, err := query.Query().WhereIn("id", []any{users[0].ID, users[1].ID}).Delete(&User{})
-					s.Equal(int64(2), res.RowsAffected)
-					s.Nil(err)
-
-					var count int64
-					s.Nil(query.Query().Model(&User{}).Where("name", "delete_user").OrWhere("name", "delete_user1").Count(&count))
-					s.True(count == 0)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestDistinct() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "distinct_user", Avatar: "distinct_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "distinct_user", Avatar: "distinct_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().Distinct("name").Find(&users, []uint{user.ID, user1.ID}))
-			s.Equal(1, len(users))
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Creating() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "trigger when create by struct",
-				setup: func() {
-					user := User{Name: "event_creating_name"}
-					s.Nil(query.Query().Create(&user))
-					s.Equal("event_creating_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_creating_name", user1.Name)
-					s.Equal("event_creating_avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create by map",
-				setup: func() {
-					s.Nil(query.Query().Model(&User{}).Create(map[string]any{
-						"name":       "event_creating_by_map_name",
-						"avatar":     "event_creating_by_map_avatar",
-						"created_at": carbon.Now(),
-						"updated_at": carbon.Now(),
-					}))
-
-					var user User
-					s.Nil(query.Query().Where("name", "event_creating_by_map_name").Find(&user))
-					s.Equal("event_creating_by_map_avatar1", user.Avatar)
-				},
-			},
-			{
-				name: "trigger when FirstOrCreate",
-				setup: func() {
-					var user User
-					s.Nil(query.Query().FirstOrCreate(&user, User{Name: "event_creating_FirstOrCreate_name"}))
-					s.True(user.ID > 0)
-					s.Equal("event_creating_FirstOrCreate_name", user.Name)
-					s.Equal("event_creating_FirstOrCreate_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_creating_FirstOrCreate_name", user1.Name)
-					s.Equal("event_creating_FirstOrCreate_avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create with omit",
-				setup: func() {
-					user := User{Name: "event_creating_omit_create_name", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "event_creating_omit_create_address"
-					user.Books[0].Name = "event_creating_omit_create_book0"
-					user.Books[1].Name = "event_creating_omit_create_book1"
-					s.Nil(query.Query().Omit("Address").Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_creating_omit_create_avatar", user.Avatar)
-					s.True(user.Address.ID == 0)
-					s.True(user.Books[0].ID > 0)
-					s.True(user.Books[1].ID > 0)
-				},
-			},
-			{
-				name: "trigger when create with select",
-				setup: func() {
-					user := User{Name: "event_creating_select_create_name", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "event_creating_select_create_address"
-					user.Books[0].Name = "event_creating_select_create_book0"
-					user.Books[1].Name = "event_creating_select_create_book1"
-					s.Nil(query.Query().Select("Name", "Avatar", "Address").Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_creating_select_create_avatar", user.Avatar)
-					s.True(user.Address.ID > 0)
-					s.True(user.Books[0].ID == 0)
-					s.True(user.Books[1].ID == 0)
-				},
-			},
-			{
-				name: "trigger when save",
-				setup: func() {
-					user := User{Name: "event_creating_save_name"}
-					s.Nil(query.Query().Save(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_creating_save_avatar", user.Avatar)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Created() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "trigger when create by struct",
-				setup: func() {
-					user := User{Name: "event_created_name", Avatar: "avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.Equal(fmt.Sprintf("event_created_avatar_%d", user.ID), user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_created_name", user1.Name)
-					s.Equal("avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create by map",
-				setup: func() {
-					userMap := map[string]any{
-						"name":       "event_created_by_map_name",
-						"avatar":     "event_created_by_map_avatar",
-						"created_at": carbon.Now(),
-						"updated_at": carbon.Now(),
-					}
-					s.Nil(query.Query().Model(&User{}).Create(userMap))
-
-					s.Equal("event_created_by_map_avatar1", userMap["avatar"])
-
-					var user User
-					s.Nil(query.Query().Where("name", "event_created_by_map_name").Find(&user))
-					s.Equal("event_created_by_map_avatar", user.Avatar)
-				},
-			},
-			{
-				name: "trigger when FirstOrCreate",
-				setup: func() {
-					var user User
-					s.Nil(query.Query().FirstOrCreate(&user, User{Name: "event_created_FirstOrCreate_name"}))
-					s.True(user.ID > 0)
-					s.Equal("event_created_FirstOrCreate_name", user.Name)
-					s.Equal(fmt.Sprintf("event_created_FirstOrCreate_avatar_%d", user.ID), user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_created_FirstOrCreate_name", user1.Name)
-					s.Empty(user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create with omit",
-				setup: func() {
-					user := User{Name: "event_created_omit_create_name", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "event_created_omit_create_address"
-					user.Books[0].Name = "event_created_omit_create_book0"
-					user.Books[1].Name = "event_created_omit_create_book1"
-					s.Nil(query.Query().Omit("Address").Create(&user))
-					s.True(user.ID > 0)
-					s.Equal(fmt.Sprintf("event_created_omit_create_avatar_%d", user.ID), user.Avatar)
-					s.True(user.Address.ID == 0)
-					s.True(user.Books[0].ID > 0)
-					s.True(user.Books[1].ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_created_omit_create_name", user1.Name)
-					s.Empty(user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create with select",
-				setup: func() {
-					user := User{Name: "event_created_select_create_name", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "event_created_select_create_address"
-					user.Books[0].Name = "event_created_select_create_book0"
-					user.Books[1].Name = "event_created_select_create_book1"
-					s.Nil(query.Query().Select("Name", "Avatar", "Address").Create(&user))
-					s.True(user.ID > 0)
-					s.Equal(fmt.Sprintf("event_created_select_create_avatar_%d", user.ID), user.Avatar)
-					s.True(user.Address.ID > 0)
-					s.True(user.Books[0].ID == 0)
-					s.True(user.Books[1].ID == 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_created_select_create_name", user1.Name)
-					s.Empty(user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when save",
-				setup: func() {
-					user := User{Name: "event_created_save_name"}
-					s.Nil(query.Query().Save(&user))
-					s.True(user.ID > 0)
-					s.Equal(fmt.Sprintf("event_created_save_avatar_%d", user.ID), user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_created_save_name", user1.Name)
-					s.Empty(user1.Avatar)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Saving() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "trigger when create by struct",
-				setup: func() {
-					user := User{Name: "event_saving_create_name"}
-					s.Nil(query.Query().Create(&user))
-					s.Equal("event_saving_create_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_saving_create_name", user1.Name)
-					s.Equal("event_saving_create_avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create by map",
-				setup: func() {
-					userMap := map[string]any{
-						"name":       "event_saving_create_by_map_name",
-						"avatar":     "event_saving_create_by_map_avatar",
-						"created_at": carbon.Now(),
-						"updated_at": carbon.Now(),
-					}
-					s.Nil(query.Query().Model(&User{}).Create(userMap))
-					s.Equal("event_saving_create_by_map_avatar1", userMap["avatar"])
-
-					var user1 User
-					s.Nil(query.Query().Where("name", "event_saving_create_by_map_name").Find(&user1))
-					s.Equal("event_saving_create_by_map_avatar1", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create with omit",
-				setup: func() {
-					user := User{Name: "event_saving_omit_create_name", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "event_saving_omit_create_address"
-					user.Books[0].Name = "event_saving_omit_create_book0"
-					user.Books[1].Name = "event_saving_omit_create_book1"
-					s.Nil(query.Query().Omit("Address").Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_saving_omit_create_avatar", user.Avatar)
-					s.True(user.Address.ID == 0)
-					s.True(user.Books[0].ID > 0)
-					s.True(user.Books[1].ID > 0)
-				},
-			},
-			{
-				name: "trigger when create with select",
-				setup: func() {
-					user := User{Name: "event_saving_select_create_name", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "event_saving_select_create_address"
-					user.Books[0].Name = "event_saving_select_create_book0"
-					user.Books[1].Name = "event_saving_select_create_book1"
-					s.Nil(query.Query().Select("Name", "Avatar", "Address").Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_saving_select_create_avatar", user.Avatar)
-					s.True(user.Address.ID > 0)
-					s.True(user.Books[0].ID == 0)
-					s.True(user.Books[1].ID == 0)
-				},
-			},
-			{
-				name: "trigger when FirstOrCreate",
-				setup: func() {
-					var user User
-					s.Nil(query.Query().FirstOrCreate(&user, User{Name: "event_saving_FirstOrCreate_name"}))
-					s.True(user.ID > 0)
-					s.Equal("event_saving_FirstOrCreate_name", user.Name)
-					s.Equal("event_saving_FirstOrCreate_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_saving_FirstOrCreate_name", user1.Name)
-					s.Equal("event_saving_FirstOrCreate_avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when save",
-				setup: func() {
-					user := User{Name: "event_saving_save_name"}
-					s.Nil(query.Query().Save(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_saving_save_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_saving_save_name", user1.Name)
-					s.Equal("event_saving_save_avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when update by single column",
-				setup: func() {
-					user := User{Name: "event_saving_single_update_name", Avatar: "avatar"}
-					s.Nil(query.Query().Create(&user))
-
-					res, err := query.Query().Model(&user).Update("avatar", "event_saving_single_update_avatar")
-					s.Nil(err)
-					s.Equal(int64(1), res.RowsAffected)
-					s.Equal("event_saving_single_update_avatar1", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_saving_single_update_name", user1.Name)
-					s.Equal("event_saving_single_update_avatar1", user1.Avatar)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Saved() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "trigger when create",
-				setup: func() {
-					user := User{Name: "event_saved_create_name", Avatar: "avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.Equal("event_saved_create_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_saved_create_name", user1.Name)
-					s.Equal("avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create by map",
-				setup: func() {
-					userMap := map[string]any{
-						"name":       "event_saved_create_by_map_name",
-						"avatar":     "event_saved_create_by_map_avatar",
-						"created_at": carbon.Now(),
-						"updated_at": carbon.Now(),
-					}
-					s.Nil(query.Query().Model(&User{}).Create(userMap))
-					s.Equal("event_saved_create_by_map_avatar1", userMap["avatar"])
-
-					var user1 User
-					s.Nil(query.Query().Where("name", "event_saved_create_by_map_name").Find(&user1))
-					s.Equal("event_saved_create_by_map_avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create with omit",
-				setup: func() {
-					user := User{Name: "event_saved_omit_create_name", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "event_saved_omit_create_address"
-					user.Books[0].Name = "event_saved_omit_create_book0"
-					user.Books[1].Name = "event_saved_omit_create_book1"
-					s.Nil(query.Query().Omit("Address").Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_saved_omit_create_avatar", user.Avatar)
-					s.True(user.Address.ID == 0)
-					s.True(user.Books[0].ID > 0)
-					s.True(user.Books[1].ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Empty(user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when create with select",
-				setup: func() {
-					user := User{Name: "event_saved_select_create_name", Address: &Address{}, Books: []*Book{{}, {}}}
-					user.Address.Name = "event_saved_select_create_address"
-					user.Books[0].Name = "event_saved_select_create_book0"
-					user.Books[1].Name = "event_saved_select_create_book1"
-					s.Nil(query.Query().Select("Name", "Avatar", "Address").Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_saved_select_create_avatar", user.Avatar)
-					s.True(user.Address.ID > 0)
-					s.True(user.Books[0].ID == 0)
-					s.True(user.Books[1].ID == 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Empty(user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when FirstOrCreate",
-				setup: func() {
-					var user User
-					s.Nil(query.Query().FirstOrCreate(&user, User{Name: "event_saved_FirstOrCreate_name"}))
-					s.True(user.ID > 0)
-					s.Equal("event_saved_FirstOrCreate_name", user.Name)
-					s.Equal("event_saved_FirstOrCreate_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_saved_FirstOrCreate_name", user1.Name)
-					s.Empty(user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when save",
-				setup: func() {
-					user := User{Name: "event_saved_save_name", Avatar: "avatar"}
-					s.Nil(query.Query().Save(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_saved_save_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_saved_save_name", user1.Name)
-					s.Equal("avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when update by map",
-				setup: func() {
-					user := User{Name: "event_saved_map_update_name", Avatar: "avatar"}
-					s.Nil(query.Query().Create(&user))
-
-					res, err := query.Query().Model(&user).Update(map[string]any{
-						"avatar": "event_saved_map_update_avatar",
-					})
-					s.Nil(err)
-					s.Equal(int64(1), res.RowsAffected)
-					s.Equal("event_saved_map_update_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_saved_map_update_name", user1.Name)
-					s.Equal("event_saved_map_update_avatar", user1.Avatar)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Updating() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "not trigger when create",
-				setup: func() {
-					user := User{Name: "event_updating_create_name", Avatar: "avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("avatar", user.Avatar)
-				},
-			},
-			{
-				name: "not trigger when create by save",
-				setup: func() {
-					user := User{Name: "event_updating_save_name", Avatar: "avatar"}
-					s.Nil(query.Query().Save(&user))
-					s.True(user.ID > 0)
-					s.Equal("avatar", user.Avatar)
-				},
-			},
-			{
-				name: "trigger when save",
-				setup: func() {
-					user := User{Name: "event_updating_save_name", Avatar: "avatar"}
-					s.Nil(query.Query().Save(&user))
-
-					user.Avatar = "event_updating_save_avatar"
-					s.Nil(query.Query().Save(&user))
-					s.Equal("event_updating_save_avatar1", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_updating_save_name", user1.Name)
-					s.Equal("event_updating_save_avatar1", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when update by model",
-				setup: func() {
-					user := User{Name: "event_updating_model_update_name", Avatar: "avatar"}
-					s.Nil(query.Query().Create(&user))
-
-					res, err := query.Query().Model(&user).Update(User{
-						Avatar: "event_updating_model_update_avatar",
-					})
-					s.Nil(err)
-					s.Equal(int64(1), res.RowsAffected)
-					s.Equal(fmt.Sprintf("event_updating_model_update_avatar_%d", user.ID), user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_updating_model_update_name", user1.Name)
-					s.Equal(fmt.Sprintf("event_updating_model_update_avatar_%d", user.ID), user1.Avatar)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Updated() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "not trigger when create",
-				setup: func() {
-					user := User{Name: "event_updated_create_name", Avatar: "avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("avatar", user.Avatar)
-				},
-			},
-			{
-				name: "not trigger when create by save",
-				setup: func() {
-					user := User{Name: "event_updated_save_name", Avatar: "avatar"}
-					s.Nil(query.Query().Save(&user))
-					s.True(user.ID > 0)
-					s.Equal("avatar", user.Avatar)
-				},
-			},
-			{
-				name: "trigger when save",
-				setup: func() {
-					user := User{Name: "event_updated_save_name", Avatar: "avatar"}
-					s.Nil(query.Query().Save(&user))
-
-					user.Avatar = "event_updated_save_avatar"
-					s.Nil(query.Query().Save(&user))
-					s.Equal("event_updated_save_avatar1", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_updated_save_name", user1.Name)
-					s.Equal("event_updated_save_avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "trigger when update by model",
-				setup: func() {
-					user := User{Name: "event_updated_model_update_name", Avatar: "avatar"}
-					s.Nil(query.Query().Create(&user))
-
-					res, err := query.Query().Model(&user).Update(User{
-						Avatar: "event_updated_model_update_avatar",
-					})
-					s.Nil(err)
-					s.Equal(int64(1), res.RowsAffected)
-					s.Equal("event_updated_model_update_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_updated_model_update_name", user1.Name)
-					s.Equal("event_updated_model_update_avatar", user1.Avatar)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Deleting() {
-	for _, query := range s.queries {
-		user := User{Name: "event_deleting_name", Avatar: "event_deleting_avatar"}
-		s.Nil(query.Query().Create(&user))
-
-		res, err := query.Query().Delete(&user)
-		s.EqualError(err, "deleting error")
-		s.Nil(res)
-
-		var user1 User
-		s.Nil(query.Query().Find(&user1, user.ID))
-		s.True(user1.ID > 0)
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Deleted() {
-	for _, query := range s.queries {
-		user := User{Name: "event_deleted_name", Avatar: "event_deleted_avatar"}
-		s.Nil(query.Query().Create(&user))
-
-		res, err := query.Query().Delete(&user)
-		s.EqualError(err, "deleted error")
-		s.Nil(res)
-
-		var user1 User
-		s.Nil(query.Query().Find(&user1, user.ID))
-		s.True(user1.ID == 0)
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_ForceDeleting() {
-	for _, query := range s.queries {
-		user := User{Name: "event_force_deleting_name", Avatar: "event_force_deleting_avatar"}
-		s.Nil(query.Query().Create(&user))
-
-		res, err := query.Query().ForceDelete(&user)
-		s.EqualError(err, "force deleting error")
-		s.Nil(res)
-
-		var user1 User
-		s.Nil(query.Query().Find(&user1, user.ID))
-		s.True(user1.ID > 0)
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_ForceDeleted() {
-	for _, query := range s.queries {
-		user := User{Name: "event_force_deleted_name", Avatar: "event_force_deleted_avatar"}
-		s.Nil(query.Query().Create(&user))
-
-		res, err := query.Query().ForceDelete(&user)
-		s.EqualError(err, "force deleted error")
-		s.Nil(res)
-
-		var user1 User
-		s.Nil(query.Query().Find(&user1, user.ID))
-		s.True(user1.ID == 0)
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Restored() {
-	for _, query := range s.queries {
-		user := User{Name: "event_restored_name", Avatar: "event_restored_avatar"}
-		s.Nil(query.Query().Create(&user))
-
-		res, err := query.Query().Delete(&user)
-		s.NoError(err)
-		s.Equal(int64(1), res.RowsAffected)
-
-		res, err = query.Query().WithTrashed().Restore(&user)
-		s.NoError(err)
-		s.Equal(int64(1), res.RowsAffected)
-		s.Equal("event_restored_name1", user.Name)
-
-		var user1 User
-		s.Nil(query.Query().Find(&user1, user.ID))
-		s.True(user1.ID > 0)
-		s.Equal("event_restored_name", user1.Name)
-		s.Equal("event_restored_avatar", user1.Avatar)
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Restoring() {
-	for _, query := range s.queries {
-		user := User{Name: "event_restoring_name", Avatar: "event_restoring_avatar"}
-		s.Nil(query.Query().Create(&user))
-
-		res, err := query.Query().Delete(&user)
-		s.NoError(err)
-		s.Equal(int64(1), res.RowsAffected)
-
-		res, err = query.Query().WithTrashed().Restore(&user)
-		s.NoError(err)
-		s.Equal(int64(1), res.RowsAffected)
-		s.Equal("event_restoring_name1", user.Name)
-
-		var user1 User
-		s.Nil(query.Query().Find(&user1, user.ID))
-		s.True(user1.ID > 0)
-		s.Equal("event_restoring_name", user1.Name)
-		s.Equal("event_restoring_avatar", user1.Avatar)
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Retrieved() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "trigger when Find",
-				setup: func() {
-					var user1 User
-					s.Nil(query.Query().Where("name", "event_retrieved_name").Find(&user1))
-					s.True(user1.ID > 0)
-					s.Equal("event_retrieved_name1", user1.Name)
-				},
-			},
-			{
-				name: "trigger when First",
-				setup: func() {
-					var user1 User
-					s.Nil(query.Query().Where("name", "event_retrieved_name").First(&user1))
-					s.True(user1.ID > 0)
-					s.Equal("event_retrieved_name1", user1.Name)
-
-					var user2 User
-					s.Nil(query.Query().Where("name", "event_retrieved_name1").First(&user2))
-					s.True(user2.ID == 0)
-					s.Empty(user2.Name)
-				},
-			},
-			{
-				name: "trigger when FirstOr",
-				setup: func() {
-					var user1 User
-					s.Nil(query.Query().Where("name", "event_retrieved_name").Find(&user1))
-					s.True(user1.ID > 0)
-					s.Equal("event_retrieved_name1", user1.Name)
-				},
-			},
-			{
-				name: "trigger when FirstOrCreate",
-				setup: func() {
-					var user1 User
-					s.Nil(query.Query().FirstOrCreate(&user1, User{Name: "event_retrieved_name"}))
-					s.True(user1.ID > 0)
-					s.Equal("event_retrieved_name1", user1.Name)
-				},
-			},
-			{
-				name: "trigger when FirstOrFail",
-				setup: func() {
-					var user1 User
-					s.Nil(query.Query().Where("name", "event_retrieved_name").FirstOrFail(&user1))
-					s.True(user1.ID > 0)
-					s.Equal("event_retrieved_name1", user1.Name)
-				},
-			},
-			{
-				name: "trigger when FirstOrNew",
-				setup: func() {
-					var user1 User
-					s.Nil(query.Query().FirstOrNew(&user1, User{Name: "event_retrieved_name"}))
-					s.True(user1.ID > 0)
-					s.Equal("event_retrieved_name1", user1.Name)
-				},
-			},
-			{
-				name: "trigger when FirstOrFail",
-				setup: func() {
-					var user1 User
-					s.Nil(query.Query().Where("name", "event_retrieved_name").FirstOrFail(&user1))
-					s.True(user1.ID > 0)
-					s.Equal("event_retrieved_name1", user1.Name)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				user := User{Name: "event_retrieved_name"}
-				s.Nil(query.Query().Create(&user))
-				s.True(user.ID > 0)
-
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_IsDirty() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "create",
-				setup: func() {
-					user := User{Name: "event_creating_IsDirty_name", Avatar: "is_dirty_avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_creating_IsDirty_avatar", user.Avatar)
-				},
-			},
-			{
-				name: "save",
-				setup: func() {
-					user := User{Name: "event_saving_IsDirty_name", Avatar: "is_dirty_avatar"}
-					s.Nil(query.Query().Save(&user))
-					s.True(user.ID > 0)
-					s.Equal("event_saving_IsDirty_avatar", user.Avatar)
-				},
-			},
-			{
-				name: "update by single column",
-				setup: func() {
-					user := User{Name: "event_updating_single_update_IsDirty_name", Avatar: "is_dirty_avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					res, err := query.Query().Model(&user).Update("name", "event_updating_single_update_IsDirty_name1")
-					s.Equal(int64(1), res.RowsAffected)
-					s.Nil(err)
-					s.Equal("event_updating_single_update_IsDirty_name1", user.Name)
-					s.Equal("event_updating_single_update_IsDirty_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_updating_single_update_IsDirty_name1", user.Name)
-					s.Equal("event_updating_single_update_IsDirty_avatar", user.Avatar)
-				},
-			},
-			{
-				name: "update by map",
-				setup: func() {
-					user := User{Name: "event_updating_map_update_IsDirty_name", Avatar: "is_dirty_avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					res, err := query.Query().Model(&user).Update(map[string]any{
-						"name": "event_updating_map_update_IsDirty_name1",
-					})
-					s.Nil(err)
-					s.Equal(int64(1), res.RowsAffected)
-					s.Equal("event_updating_map_update_IsDirty_name1", user.Name)
-					s.Equal("event_updating_map_update_IsDirty_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_updating_map_update_IsDirty_name1", user.Name)
-					s.Equal("event_updating_map_update_IsDirty_avatar", user.Avatar)
-				},
-			},
-			{
-				name: "update by model",
-				setup: func() {
-					user := User{Name: "event_updating_model_update_IsDirty_name", Avatar: "is_dirty_avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					res, err := query.Query().Model(&user).Update(User{
-						Name: "event_updating_model_update_IsDirty_name1",
-					})
-					s.Equal(int64(1), res.RowsAffected)
-					s.Nil(err)
-					s.Equal("event_updating_model_update_IsDirty_name1", user.Name)
-					s.Equal("event_updating_model_update_IsDirty_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_updating_model_update_IsDirty_name1", user.Name)
-					s.Equal("event_updating_model_update_IsDirty_avatar", user.Avatar)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Context() {
-	for _, query := range s.queries {
-		user := User{Name: "event_context"}
-		s.Nil(query.Query().Create(&user))
-		s.Equal("goravel", user.Avatar)
-	}
-}
-
-func (s *OrmTestSuite) TestEvent_Query() {
-	for _, query := range s.queries {
-		user := User{Name: "event_query"}
-		s.Nil(query.Query().Create(&user))
-		s.True(user.ID > 0)
-		s.Equal("event_query", user.Name)
-
-		var user1 User
-		s.Nil(query.Query().Where("name", "event_query1").Find(&user1))
-		s.True(user1.ID > 0)
-	}
-}
-
-func (s *OrmTestSuite) TestExec() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			res, err := query.Query().Exec("INSERT INTO users (name, avatar, created_at, updated_at) VALUES ('exec_user', 'exec_avatar', '2023-03-09 18:56:33', '2023-03-09 18:56:35');")
-			s.Equal(int64(1), res.RowsAffected)
-			s.Nil(err)
-
-			var user User
-			err = query.Query().Where("name", "exec_user").First(&user)
-			s.Nil(err)
-			s.True(user.ID > 0)
-
-			res, err = query.Query().Exec(fmt.Sprintf("UPDATE users set name = 'exec_user1' where id = %d", user.ID))
-			s.Equal(int64(1), res.RowsAffected)
-			s.Nil(err)
-
-			res, err = query.Query().Exec(fmt.Sprintf("DELETE FROM users where id = %d", user.ID))
-			s.Equal(int64(1), res.RowsAffected)
-			s.Nil(err)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestExists() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "exists_user", Avatar: "exists_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "exists_user", Avatar: "exists_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var t bool
-			s.Nil(query.Query().Model(&User{}).Where("name = ?", "exists_user").Exists(&t))
-			s.True(t)
-
-			var f bool
-			s.Nil(query.Query().Model(&User{}).Where("name = ?", "no_exists_user").Exists(&f))
-			s.False(f)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestFind() {
-	for _, query := range s.queries {
-		user := User{Name: "find_user"}
-		s.Nil(query.Query().Create(&user))
-		s.True(user.ID > 0)
-
-		var user2 User
-		s.Nil(query.Query().Find(&user2, user.ID))
-		s.True(user2.ID > 0)
-
-		var user3 []User
-		s.Nil(query.Query().Find(&user3, []uint{user.ID}))
-		s.Equal(1, len(user3))
-
-		var user4 []User
-		s.Nil(query.Query().Where("id in ?", []uint{user.ID}).Find(&user4))
-		s.Equal(1, len(user4))
-	}
-}
-
-func (s *OrmTestSuite) TestFindOrFail() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "success",
-				setup: func() {
-					user := User{Name: "find_user"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					var user2 User
-					s.Nil(query.Query().FindOrFail(&user2, user.ID))
-					s.True(user2.ID > 0)
-				},
-			},
-			{
-				name: "error",
-				setup: func() {
-					var user User
-					s.ErrorIs(query.Query().FindOrFail(&user, 10000), errors.OrmRecordNotFound)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-// func (s *OrmTestSuite) TestFirst() {
-// 	for _, query := range s.queries {
-// 		user := User{Name: "first_user"}
-// 		s.Nil(query.Query().Create(&user))
-// 		s.True(user.ID > 0)
-
-// 		var user1 User
-// 		s.Nil(query.Query().Where("name", "first_user").First(&user1))
-// 		s.True(user1.ID > 0)
-
-// 		// refresh connection
-// 		mockCommonConnection(query.MockConfig(), query, "dummy")
-
-// 		people := People{Body: "first_people"}
-// 		s.Nil(query.Query().Create(&people))
-// 		s.True(people.ID > 0)
-
-// 		var people1 People
-// 		s.Nil(query.Query().Where("id in ?", []uint{people.ID}).First(&people1))
-// 		s.True(people1.ID > 0)
-// 	}
-// }
-
-func (s *OrmTestSuite) TestFirstOr() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "not found, new one",
-				setup: func() {
-					var user User
-					s.Nil(query.Query().Where("name", "first_or_user").FirstOr(&user, func() error {
-						user.Name = "goravel"
-
-						return nil
-					}))
-					s.Equal(uint(0), user.ID)
-					s.Equal("goravel", user.Name)
-
-				},
-			},
-			{
-				name: "found",
-				setup: func() {
-					user := User{Name: "first_or_name"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Where("name", "first_or_name").Find(&user1))
-					s.True(user1.ID > 0)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestFirstOrCreate() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "error when empty conditions",
-				setup: func() {
-					var user User
-					s.EqualError(query.Query().FirstOrCreate(&user), errors.OrmQueryConditionRequired.Error())
-					s.True(user.ID == 0)
-				},
-			},
-			{
-				name: "success",
-				setup: func() {
-					var user User
-					s.Nil(query.Query().FirstOrCreate(&user, User{Name: "first_or_create_user"}))
-					s.True(user.ID > 0)
-					s.Equal("first_or_create_user", user.Name)
-
-					var user1 User
-					s.Nil(query.Query().FirstOrCreate(&user1, User{Name: "first_or_create_user"}))
-					s.Equal(user.ID, user1.ID)
-
-					var user2 User
-					s.Nil(query.Query().Where("avatar", "first_or_create_avatar").FirstOrCreate(&user2, User{Name: "user"}, User{Avatar: "first_or_create_avatar2"}))
-					s.True(user2.ID > 0)
-					s.True(user2.Avatar == "first_or_create_avatar2")
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestFirstOrFail() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "fail",
-				setup: func() {
-					var user User
-					s.ErrorIs(query.Query().Where("name", "first_or_fail_user").FirstOrFail(&user), errors.OrmRecordNotFound)
-					s.Equal(uint(0), user.ID)
-				},
-			},
-			{
-				name: "success",
-				setup: func() {
-					user := User{Name: "first_or_fail_name"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("first_or_fail_name", user.Name)
-
-					var user1 User
-					s.Nil(query.Query().Where("name", "first_or_fail_name").FirstOrFail(&user1))
-					s.True(user1.ID > 0)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestFirstOrNew() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "not found, new one",
-				setup: func() {
-					var user User
-					s.Nil(query.Query().FirstOrNew(&user, User{Name: "first_or_new_name"}))
-					s.Equal(uint(0), user.ID)
-					s.Equal("first_or_new_name", user.Name)
-					s.Empty(user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().FirstOrNew(&user1, User{Name: "first_or_new_name"}, User{Avatar: "first_or_new_avatar"}))
-					s.Equal(uint(0), user1.ID)
-					s.Equal("first_or_new_name", user1.Name)
-					s.Equal("first_or_new_avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "found",
-				setup: func() {
-					user := User{Name: "first_or_new_name"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("first_or_new_name", user.Name)
-
-					var user1 User
-					s.Nil(query.Query().FirstOrNew(&user1, User{Name: "first_or_new_name"}))
-					s.True(user1.ID > 0)
-					s.Equal("first_or_new_name", user1.Name)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestForceDelete() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "success",
-				setup: func() {
-					user := User{Name: "force_delete_name"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("force_delete_name", user.Name)
-
-					res, err := query.Query().Where("name", "force_delete_name").ForceDelete(&User{})
-					s.Equal(int64(1), res.RowsAffected)
-					s.Nil(err)
-					s.Equal("force_delete_name", user.Name)
-
-					var user1 User
-					s.Nil(query.Query().WithTrashed().Find(&user1, user.ID))
-					s.Equal(uint(0), user1.ID)
-				},
-			},
-			{
-				name: "success by table",
-				setup: func() {
-					user := User{Name: "force_delete_name_by_table"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("force_delete_name_by_table", user.Name)
-
-					res, err := query.Query().Table("users").Where("name", "force_delete_name_by_table").ForceDelete()
-					s.Equal(int64(1), res.RowsAffected)
-					s.Nil(err)
-					s.Equal("force_delete_name_by_table", user.Name)
-
-					var user1 User
-					s.Nil(query.Query().WithTrashed().Find(&user1, user.ID))
-					s.Equal(uint(0), user1.ID)
-				},
-			},
-			{
-				name: "success by model",
-				setup: func() {
-					user := User{Name: "force_delete_name_by_model"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-					s.Equal("force_delete_name_by_model", user.Name)
-
-					res, err := query.Query().Model(&User{}).Where("name", "force_delete_name_by_model").ForceDelete()
-					s.Equal(int64(1), res.RowsAffected)
-					s.Nil(err)
-					s.Equal("force_delete_name_by_model", user.Name)
-
-					var user1 User
-					s.Nil(query.Query().WithTrashed().Find(&user1, user.ID))
-					s.Equal(uint(0), user1.ID)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-// func (s *OrmTestSuite) TestGet() {
-// 	for driver, query := range s.queries {
-// 		s.Run(driver, func() {
-// 			user := User{Name: "get_user"}
-// 			s.Nil(query.Query().Create(&user))
-// 			s.True(user.ID > 0)
-
-// 			var user1 []User
-// 			s.Nil(query.Query().Where("id in ?", []uint{user.ID}).Get(&user1))
-// 			s.Equal(1, len(user1))
-
-// 			// refresh connection
-// 			mockCommonConnection(query.MockConfig(), query, "dummy")
-
-// 			people := People{Body: "get_people"}
-// 			s.Nil(query.Query().Create(&people))
-// 			s.True(people.ID > 0)
-
-// 			var people1 []People
-// 			s.Nil(query.Query().Where("id in ?", []uint{people.ID}).Get(&people1))
-// 			s.Equal(1, len(people1))
-
-// 			var user2 []User
-// 			s.Nil(query.Query().Where("id in ?", []uint{user.ID}).Get(&user2))
-// 			s.Equal(1, len(user2))
-// 		})
-// 	}
-// }
-
-func (s *OrmTestSuite) TestJoin() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "join_user", Avatar: "join_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			userAddress := Address{UserID: user.ID, Name: "join_address", Province: "join_province"}
-			s.Nil(query.Query().Create(&userAddress))
-			s.True(userAddress.ID > 0)
-
-			type Result struct {
-				UserName        string
-				UserAddressName string
-			}
-			var result []Result
-			s.Nil(query.Query().Model(&User{}).Where("users.id = ?", user.ID).Join("left join addresses ua on users.id = ua.user_id").
-				Select("users.name user_name, ua.name user_address_name").Get(&result))
-			s.Equal(1, len(result))
-			s.Equal("join_user", result[0].UserName)
-			s.Equal("join_address", result[0].UserAddressName)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestLockForUpdate() {
-	for driver, query := range s.queries {
-		if driver != DriverSqlite {
-			s.Run(driver, func() {
-				user := User{Name: "lock_for_update_user"}
-				s.Nil(query.Query().Create(&user))
-				s.True(user.ID > 0)
-
-				for i := 0; i < 10; i++ {
-					go func() {
-						tx, err := query.Query().Begin()
-						s.Nil(err)
-
-						var user1 User
-						s.Nil(tx.LockForUpdate().Find(&user1, user.ID))
-						s.True(user1.ID > 0)
-						user1.Name += "1"
-						s.Nil(tx.Save(&user1))
-
-						s.Nil(tx.Commit())
-					}()
-				}
-
-				time.Sleep(2 * time.Second)
-
-				var user2 User
-				s.Nil(query.Query().Find(&user2, user.ID))
-				s.Equal("lock_for_update_user1111111111", user2.Name)
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestOffset() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "offset_user", Avatar: "offset_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "offset_user", Avatar: "offset_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var user2 []User
-			s.Nil(query.Query().Where("name = ?", "offset_user").Offset(1).Limit(1).Get(&user2))
-			s.True(len(user2) > 0)
-			s.True(user2[0].ID > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestOrder() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "order_user", Avatar: "order_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "order_user", Avatar: "order_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var user2 []User
-			s.Nil(query.Query().Where("name = ?", "order_user").Order("id desc").Order("name asc").Get(&user2))
-			s.True(len(user2) > 0)
-			s.True(user2[0].ID > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestOrderBy() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "order_asc_user", Avatar: "order_asc_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "order_asc_user", Avatar: "order_asc_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var users1 []User
-			s.Nil(query.Query().Where("name = ?", "order_asc_user").OrderBy("id").Get(&users1))
-			s.True(len(users1) == 2)
-			s.True(users1[0].ID == user.ID)
-
-			var users2 []User
-			s.Nil(query.Query().Where("name = ?", "order_asc_user").OrderBy("id", "DESC").Get(&users2))
-			s.True(len(users2) == 2)
-			s.True(users2[0].ID == user1.ID)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestOrderByDesc() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "order_desc_user", Avatar: "order_desc_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "order_desc_user", Avatar: "order_desc_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().Where("name = ?", "order_desc_user").OrderByDesc("id").Get(&users))
-			usersLength := len(users)
-			s.True(usersLength == 2)
-			s.True(users[usersLength-1].ID == user.ID)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestInRandomOrder() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			for i := 0; i < 30; i++ {
-				user := User{Name: "random_order_user", Avatar: "random_order_avatar"}
-				s.Nil(query.Query().Create(&user))
-				s.True(user.ID > 0)
-			}
-
-			var users1 []User
-			s.Nil(query.Query().Where("name = ?", "random_order_user").InRandomOrder().Find(&users1))
-			s.True(len(users1) == 30)
-
-			var users2 []User
-			s.Nil(query.Query().Where("name = ?", "random_order_user").InRandomOrder().Find(&users2))
-			s.True(len(users2) == 30)
-
-			s.True(users1[0].ID != users2[0].ID || users1[14].ID != users2[14].ID || users1[29].ID != users2[29].ID)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestInTransaction() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			s.False(query.Query().InTransaction())
-
-			tx, err := query.Query().Begin()
-			s.NotNil(tx)
-			s.NoError(err)
-
-			s.True(tx.InTransaction())
-			s.NoError(tx.Commit())
-			s.False(query.Query().InTransaction())
-
-			tx, err = query.Query().Begin()
-			s.NotNil(tx)
-			s.NoError(err)
-
-			s.True(tx.InTransaction())
-			s.NoError(tx.Rollback())
-			s.False(query.Query().InTransaction())
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestPaginate() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "paginate_user", Avatar: "paginate_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "paginate_user", Avatar: "paginate_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			user2 := User{Name: "paginate_user", Avatar: "paginate_avatar2"}
-			s.Nil(query.Query().Create(&user2))
-			s.True(user2.ID > 0)
-
-			user3 := User{Name: "paginate_user", Avatar: "paginate_avatar3"}
-			s.Nil(query.Query().Create(&user3))
-			s.True(user3.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().Where("name = ?", "paginate_user").Paginate(1, 3, &users, nil))
-			s.Equal(3, len(users))
-
-			var users1 []User
-			var total1 int64
-			s.Nil(query.Query().Where("name = ?", "paginate_user").Paginate(2, 3, &users1, &total1))
-			s.Equal(1, len(users1))
-			s.Equal(int64(4), total1)
-
-			var users2 []User
-			var total2 int64
-			s.Nil(query.Query().Model(User{}).Where("name = ?", "paginate_user").Paginate(1, 3, &users2, &total2))
-			s.Equal(3, len(users2))
-			s.Equal(int64(4), total2)
-
-			var users3 []User
-			var total3 int64
-			s.Nil(query.Query().Table("users").Where("name = ?", "paginate_user").Paginate(1, 3, &users3, &total3))
-			s.Equal(3, len(users3))
-			s.Equal(int64(4), total3)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestPluck() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "pluck_user", Avatar: "pluck_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "pluck_user", Avatar: "pluck_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var avatars []string
-			s.Nil(query.Query().Model(&User{}).Where("name = ?", "pluck_user").Pluck("avatar", &avatars))
-			s.Equal(2, len(avatars))
-			s.Equal("pluck_avatar", avatars[0])
-			s.Equal("pluck_avatar1", avatars[1])
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestHasOne() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := &User{
-				Name: "has_one_name",
-				Address: &Address{
-					Name: "has_one_address",
-				},
-			}
-
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-			s.True(user.Address.ID > 0)
-
-			var user1 User
-			s.Nil(query.Query().With("Address").Where("name = ?", "has_one_name").First(&user1))
-			s.True(user.ID > 0)
-			s.True(user.Address.ID > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestHasOneMorph() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := &User{
-				Name: "has_one_morph_name",
-				House: &House{
-					Name: "has_one_morph_house",
-				},
-			}
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-			s.True(user.House.ID > 0)
-
-			var user1 User
-			s.Nil(query.Query().With("House").Where("name = ?", "has_one_morph_name").First(&user1))
-			s.True(user.ID > 0)
-			s.True(user.Name == "has_one_morph_name")
-			s.True(user.House.ID > 0)
-			s.True(user.House.Name == "has_one_morph_house")
-
-			var house House
-			s.Nil(query.Query().Where("name = ?", "has_one_morph_house").Where("houseable_type = ?", "users").Where("houseable_id = ?", user.ID).First(&house))
-			s.True(house.ID > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestHasMany() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := &User{
-				Name: "has_many_name",
-				Books: []*Book{
-					{Name: "has_many_book1"},
-					{Name: "has_many_book2"},
-				},
-			}
-
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-			s.True(user.Books[0].ID > 0)
-			s.True(user.Books[1].ID > 0)
-
-			var user1 User
-			s.Nil(query.Query().With("Books").Where("name = ?", "has_many_name").First(&user1))
-			s.True(user.ID > 0)
-			s.True(len(user.Books) == 2)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestHasManyMorph() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := &User{
-				Name: "has_many_morph_name",
-				Phones: []*Phone{
-					{Name: "has_many_morph_phone1"},
-					{Name: "has_many_morph_phone2"},
-				},
-			}
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-			s.True(user.Phones[0].ID > 0)
-			s.True(user.Phones[1].ID > 0)
-
-			var user1 User
-			s.Nil(query.Query().With("Phones").Where("name = ?", "has_many_morph_name").First(&user1))
-			s.True(user.ID > 0)
-			s.True(user.Name == "has_many_morph_name")
-			s.True(len(user.Phones) == 2)
-			s.True(user.Phones[0].Name == "has_many_morph_phone1")
-			s.True(user.Phones[1].Name == "has_many_morph_phone2")
-
-			var phones []Phone
-			s.Nil(query.Query().Where("name like ?", "has_many_morph_phone%").Where("phoneable_type = ?", "users").Where("phoneable_id = ?", user.ID).Find(&phones))
-			s.True(len(phones) == 2)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestManyToMany() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := &User{
-				Name: "many_to_many_name",
-				Roles: []*Role{
-					{Name: "many_to_many_role1"},
-					{Name: "many_to_many_role2"},
-				},
-			}
-
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-			s.True(user.Roles[0].ID > 0)
-			s.True(user.Roles[1].ID > 0)
-
-			var user1 User
-			s.Nil(query.Query().With("Roles").Where("name = ?", "many_to_many_name").First(&user1))
-			s.True(user.ID > 0)
-			s.True(len(user.Roles) == 2)
-
-			var role Role
-			s.Nil(query.Query().With("Users").Where("name = ?", "many_to_many_role1").First(&role))
-			s.True(role.ID > 0)
-			s.True(len(role.Users) == 1)
-			s.Equal("many_to_many_name", role.Users[0].Name)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestLimit() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "limit_user", Avatar: "limit_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "limit_user", Avatar: "limit_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var user2 []User
-			s.Nil(query.Query().Where("name = ?", "limit_user").Limit(1).Get(&user2))
-			s.True(len(user2) > 0)
-			s.True(user2[0].ID > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestLoad() {
-	for _, query := range s.queries {
-		user := User{Name: "load_user", Address: &Address{}, Books: []*Book{{}, {}}}
-		user.Address.Name = "load_address"
-		user.Books[0].Name = "load_book0"
-		user.Books[1].Name = "load_book1"
-		s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-		s.True(user.ID > 0)
-		s.True(user.Address.ID > 0)
-		s.True(user.Books[0].ID > 0)
-		s.True(user.Books[1].ID > 0)
-
-		tests := []struct {
-			description string
-			setup       func(description string)
-		}{
-			{
-				description: "simple load relationship",
-				setup: func(description string) {
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Nil(user1.Address)
-					s.True(len(user1.Books) == 0)
-					s.Nil(query.Query().Load(&user1, "Address"))
-					s.True(user1.Address.ID > 0)
-					s.True(len(user1.Books) == 0)
-					s.Nil(query.Query().Load(&user1, "Books"))
-					s.True(user1.Address.ID > 0)
-					s.True(len(user1.Books) == 2)
-				},
-			},
-			{
-				description: "load relationship with simple condition",
-				setup: func(description string) {
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Nil(user1.Address)
-					s.Equal(0, len(user1.Books))
-					s.Nil(query.Query().Load(&user1, "Books", "name = ?", "load_book0"))
-					s.True(user1.ID > 0)
-					s.Nil(user1.Address)
-					s.Equal(1, len(user1.Books))
-					s.Equal("load_book0", user.Books[0].Name)
-				},
-			},
-			{
-				description: "load relationship with func condition",
-				setup: func(description string) {
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Nil(user1.Address)
-					s.Equal(0, len(user1.Books))
-					s.Nil(query.Query().Load(&user1, "Books", func(query contractsorm.Query) contractsorm.Query {
-						return query.Where("name = ?", "load_book0")
-					}))
-					s.True(user1.ID > 0)
-					s.Nil(user1.Address)
-					s.Equal(1, len(user1.Books))
-					s.Equal("load_book0", user.Books[0].Name)
-				},
-			},
-			{
-				description: "error when relation is empty",
-				setup: func(description string) {
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.True(user1.ID > 0)
-					s.Nil(user1.Address)
-					s.Equal(0, len(user1.Books))
-					s.EqualError(query.Query().Load(&user1, ""), errors.OrmQueryEmptyRelation.Error())
-				},
-			},
-			{
-				description: "error when id is nil",
-				setup: func(description string) {
-					type UserNoID struct {
-						Name   string
-						Avatar string
-					}
-					var userNoID UserNoID
-					s.EqualError(query.Query().Load(&userNoID, "Book"), errors.OrmQueryEmptyId.Error())
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.description, func() {
-				test.setup(test.description)
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestLoadMissing() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "load_missing_user", Address: &Address{}, Books: []*Book{{}, {}}}
-			user.Address.Name = "load_missing_address"
-			user.Books[0].Name = "load_missing_book0"
-			user.Books[1].Name = "load_missing_book1"
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-			s.True(user.Address.ID > 0)
-			s.True(user.Books[0].ID > 0)
-			s.True(user.Books[1].ID > 0)
-
-			tests := []struct {
-				description string
-				setup       func(description string)
-			}{
-				{
-					description: "load when missing",
-					setup: func(description string) {
-						var user1 User
-						s.Nil(query.Query().Find(&user1, user.ID))
-						s.True(user1.ID > 0)
-						s.Nil(user1.Address)
-						s.True(len(user1.Books) == 0)
-						s.Nil(query.Query().LoadMissing(&user1, "Address"))
-						s.True(user1.Address.ID > 0)
-						s.True(len(user1.Books) == 0)
-						s.Nil(query.Query().LoadMissing(&user1, "Books"))
-						s.True(user1.Address.ID > 0)
-						s.True(len(user1.Books) == 2)
-					},
-				},
-				{
-					description: "don't load when not missing",
-					setup: func(description string) {
-						var user1 User
-						s.Nil(query.Query().With("Books", "name = ?", "load_missing_book0").Find(&user1, user.ID))
-						s.True(user1.ID > 0)
-						s.Nil(user1.Address)
-						s.True(len(user1.Books) == 1)
-						s.Nil(query.Query().LoadMissing(&user1, "Address"))
-						s.True(user1.Address.ID > 0)
-						s.Nil(query.Query().LoadMissing(&user1, "Books"))
-						s.True(len(user1.Books) == 1)
-					},
-				},
-			}
-			for _, test := range tests {
-				test.setup(test.description)
-			}
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestModel() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			// model is valid
-			user := User{Name: "model_user"}
-			s.Nil(query.Query().Model(&User{}).Create(&user))
-			s.True(user.ID > 0)
-
-			// model is invalid
-			user1 := User{Name: "model_user"}
-			s.EqualError(query.Query().Model("users").Create(&user1), errors.OrmQueryInvalidModel.Args("").Error())
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestRaw() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "raw_user", Avatar: "raw_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			var user1 User
-			s.Nil(query.Query().Raw("SELECT id, name FROM users WHERE name = ?", "raw_user").Scan(&user1))
-			s.True(user1.ID > 0)
-			s.Equal("raw_user", user1.Name)
-			s.Empty(user1.Avatar)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestReuse() {
-	for _, query := range s.queries {
-		users := []User{{Name: "reuse_user", Avatar: "reuse_avatar"}, {Name: "reuse_user1", Avatar: "reuse_avatar1"}}
-		s.Nil(query.Query().Create(&users))
-		s.True(users[0].ID > 0)
-		s.True(users[1].ID > 0)
-
-		q := query.Query().Where("name", "reuse_user")
-
-		var users1 User
-		s.Nil(q.Where("avatar", "reuse_avatar").Find(&users1))
-		s.True(users1.ID > 0)
-
-		var users2 User
-		s.Nil(q.Where("avatar", "reuse_avatar1").Find(&users2))
-		s.True(users2.ID == 0)
-
-		var users3 User
-		s.Nil(query.Query().Where("avatar", "reuse_avatar1").Find(&users3))
-		s.True(users3.ID > 0)
-	}
-}
-
-// func (s *OrmTestSuite) TestRefreshConnection() {
-// 	tests := []struct {
-// 		name             string
-// 		model            any
-// 		setup            func()
-// 		expectConnection string
-// 		expectErr        string
-// 	}{
-// 		{
-// 			name: "invalid model",
-// 			model: func() any {
-// 				var product string
-// 				return product
-// 			}(),
-// 			setup:     func() {},
-// 			expectErr: errors.OrmQueryInvalidModel.Args("").Error(),
-// 		},
-// 		{
-// 			name: "the connection of model is empty",
-// 			model: func() any {
-// 				var review Review
-// 				return review
-// 			}(),
-// 			setup:            func() {},
-// 			expectConnection: "postgres",
-// 		},
-// 		{
-// 			name: "the connection of model is same as current connection",
-// 			model: func() any {
-// 				var box Box
-// 				return box
-// 			}(),
-// 			setup:            func() {},
-// 			expectConnection: "postgres",
-// 		},
-// 		{
-// 			name: "connections are different, but drivers are same",
-// 			model: func() any {
-// 				var people People
-// 				return people
-// 			}(),
-// 			setup: func() {
-// 				mockCommonConnection(s.queries[database.DriverPostgres].MockConfig(), s.additionalQuery, "dummy")
-// 			},
-// 			expectConnection: "dummy",
-// 		},
-// 		{
-// 			name: "connections and drivers are different",
-// 			model: func() any {
-// 				var product Product
-// 				return product
-// 			}(),
-// 			setup: func() {
-// 				mockCommonConnection(s.queries[database.DriverPostgres].MockConfig(), s.queries[database.DriverSqlite], "sqlite")
-// 			},
-// 			expectConnection: "sqlite",
-// 		},
-// 	}
-
-// 	for _, test := range tests {
-// 		s.Run(test.name, func() {
-// 			test.setup()
-// 			testQuery := s.queries[database.DriverPostgres]
-// 			query, err := testQuery.Query().(*Query).refreshConnection(test.model)
-// 			if test.expectErr != "" {
-// 				s.EqualError(err, test.expectErr)
-// 			} else {
-// 				s.Nil(err)
-// 			}
-// 			if test.expectConnection == "" {
-// 				s.Nil(query)
-// 			} else {
-// 				s.Equal(test.expectConnection, query.dbConfig.Connection)
-// 			}
-// 		})
-// 	}
-// }
-
-func (s *OrmTestSuite) TestSave() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "success when create",
-				setup: func() {
-					user := User{
-						Name:   "save_create_user",
-						Avatar: "save_create_avatar",
-						House:  &House{Name: "save_create_house"},
-					}
-					s.Nil(query.Query().Save(&user))
-					s.True(user.ID > 0)
-					s.True(user.House.ID == 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("save_create_user", user1.Name)
-				},
-			},
-			{
-				name: "success when create with Select",
-				setup: func() {
-					user := User{
-						Name:   "save_create_with_select_user",
-						Avatar: "save_create_with_select_avatar",
-						House:  &House{Name: "save_create_with_select_house"},
-					}
-					s.Nil(query.Query().Select("Name", "House").Save(&user))
-					s.True(user.ID > 0)
-					s.True(user.House.ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("save_create_with_select_user", user1.Name)
-					s.Empty(user1.Avatar)
-
-					var house1 House
-					s.Nil(query.Query().Find(&house1, user.House.ID))
-					s.Equal("save_create_with_select_house", house1.Name)
-				},
-			},
-			{
-				name: "success when create with Omit",
-				setup: func() {
-					user := User{
-						Name:   "save_create_with_omit_user",
-						Avatar: "save_create_with_omit_avatar",
-						House:  &House{Name: "save_create_with_omit_house"},
-					}
-					s.Nil(query.Query().Omit("House").Save(&user))
-					s.True(user.ID > 0)
-					s.True(user.House.ID == 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("save_create_with_omit_user", user1.Name)
-					s.Equal("save_create_with_omit_avatar", user1.Avatar)
-				},
-			},
-			{
-				name: "success when update",
-				setup: func() {
-					user := User{Name: "save_update_user", Avatar: "save_update_avatar"}
-					s.Nil(query.Query().Create(&user))
-					s.True(user.ID > 0)
-
-					user.Name = "save_update_user1"
-					s.Nil(query.Query().Save(&user))
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("save_update_user1", user1.Name)
-				},
-			},
-			{
-				name: "success when update with Select",
-				setup: func() {
-					user := User{
-						Name:   "save_update_with_select_user",
-						Avatar: "save_update_with_select_avatar",
-						House:  &House{Name: "save_update_with_select_house"},
-					}
-					s.Nil(query.Query().Select("Name", "Avatar").Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.House.ID == 0)
-
-					user.Name = "save_update_with_select_user1"
-					s.Nil(query.Query().Select("Name", "House").Save(&user))
-
-					s.True(user.House.ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("save_update_with_select_user1", user1.Name)
-
-					var house1 House
-					s.Nil(query.Query().Find(&house1, user.House.ID))
-					s.Equal("save_update_with_select_house", house1.Name)
-				},
-			},
-			{
-				name: "success when update with Omit",
-				setup: func() {
-					user := User{
-						Name:    "save_update_with_omit_user",
-						Avatar:  "save_update_with_omit_avatar",
-						Address: &Address{Name: "save_update_with_omit_address"},
-						House:   &House{Name: "save_update_with_omit_house"},
-					}
-					s.Nil(query.Query().Select("Name", "Avatar", "Address").Create(&user))
-					s.True(user.ID > 0)
-					s.True(user.Address.ID > 0)
-					s.True(user.House.ID == 0)
-
-					user.Name = "save_update_with_omit_user1"
-					user.Address.Name = "save_update_with_omit_address1"
-					s.Nil(query.Query().Omit("Address").Save(&user))
-
-					s.True(user.House.ID > 0)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("save_update_with_omit_user1", user1.Name)
-
-					var address1 Address
-					s.Nil(query.Query().Find(&address1, user.Address.ID))
-					s.Equal("save_update_with_omit_address", address1.Name)
-
-					var house1 House
-					s.Nil(query.Query().Find(&house1, user.House.ID))
-					s.Equal("save_update_with_omit_house", house1.Name)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestSaveQuietly() {
-	for _, query := range s.queries {
-		user := User{Name: "event_save_quietly_name", Avatar: "save_quietly_avatar"}
-		s.Nil(query.Query().SaveQuietly(&user))
-		s.True(user.ID > 0)
-		s.Equal("event_save_quietly_name", user.Name)
-		s.Equal("save_quietly_avatar", user.Avatar)
-
-		var user1 User
-		s.Nil(query.Query().Find(&user1, user.ID))
-		s.Equal("event_save_quietly_name", user1.Name)
-		s.Equal("save_quietly_avatar", user1.Avatar)
-	}
-}
-
-func (s *OrmTestSuite) TestScope() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			users := []User{{Name: "scope_user", Avatar: "scope_avatar"}, {Name: "scope_user1", Avatar: "scope_avatar1"}}
-			s.Nil(query.Query().Create(&users))
-			s.True(users[0].ID > 0)
-			s.True(users[1].ID > 0)
-
-			var users1 []User
-			s.Nil(query.Query().Scopes(paginator("1", "1")).Find(&users1))
-
-			s.Equal(1, len(users1))
-			s.True(users1[0].ID > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestSelect() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "select_user", Avatar: "select_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "select_user", Avatar: "select_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			user2 := User{Name: "select_user1", Avatar: "select_avatar1"}
-			s.Nil(query.Query().Create(&user2))
-			s.True(user2.ID > 0)
-
-			type Result struct {
-				Name  string
-				Count string
-			}
-			var result []Result
-			s.Nil(query.Query().Model(&User{}).Select("name, count(avatar) as count").Where("id in ?", []uint{user.ID, user1.ID, user2.ID}).Group("name").Get(&result))
-			s.Equal(2, len(result))
-			s.Equal("select_user", result[0].Name)
-			s.Equal("2", result[0].Count)
-			s.Equal("select_user1", result[1].Name)
-			s.Equal("1", result[1].Count)
-
-			var result1 []Result
-			s.Nil(query.Query().Model(&User{}).Select("name, count(avatar) as count").Group("name").Having("name = ?", "select_user").Get(&result1))
-
-			s.Equal(1, len(result1))
-			s.Equal("select_user", result1[0].Name)
-			s.Equal("2", result1[0].Count)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestSharedLock() {
-	for driver, query := range s.queries {
-		if driver != DriverSqlite {
-			s.Run(driver, func() {
-				user := User{Name: "shared_lock_user"}
-				s.Nil(query.Query().Create(&user))
-				s.True(user.ID > 0)
-
-				tx, err := query.Query().Begin()
-				s.Nil(err)
-				var user1 User
-				s.Nil(tx.SharedLock().Find(&user1, user.ID))
-				s.True(user1.ID > 0)
-
-				var user2 User
-				s.Nil(query.Query().SharedLock().Find(&user2, user.ID))
-				s.True(user2.ID > 0)
-
-				user1.Name += "1"
-				s.Nil(tx.Save(&user1))
-
-				s.Nil(tx.Commit())
-
-				var user3 User
-				s.Nil(query.Query().Find(&user3, user.ID))
-				s.Equal("shared_lock_user1", user3.Name)
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestSoftDelete() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "soft_delete_user", Avatar: "soft_delete_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			res, err := query.Query().Where("name = ?", "soft_delete_user").Delete(&User{})
-			s.Equal(int64(1), res.RowsAffected)
-			s.Nil(err)
-
-			var user1 User
-			s.Nil(query.Query().Find(&user1, user.ID))
-			s.Equal(uint(0), user1.ID)
-
-			var user2 User
-			s.Nil(query.Query().WithTrashed().Find(&user2, user.ID))
-			s.True(user2.ID > 0)
-
-			res, err = query.Query().Where("name = ?", "soft_delete_user").ForceDelete(&User{})
-			s.Equal(int64(1), res.RowsAffected)
-			s.Nil(err)
-
-			var user3 User
-			s.Nil(query.Query().WithTrashed().Find(&user3, user.ID))
-			s.Equal(uint(0), user3.ID)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestRestore() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			users := []User{
-				{Name: "restore_user1", Avatar: "restore_avatar"},
-				{Name: "restore_user2", Avatar: "restore_avatar"},
-				{Name: "restore_user3", Avatar: "restore_avatar"},
-				{Name: "restore_user4", Avatar: "restore_avatar"},
-			}
-			s.NoError(query.Query().Create(&users))
-			s.True(users[0].ID > 0)
-			s.True(users[1].ID > 0)
-			s.True(users[2].ID > 0)
-			s.True(users[3].ID > 0)
-
-			res, err := query.Query().Where("avatar", "restore_avatar").Delete(&User{})
-			s.Equal(int64(4), res.RowsAffected)
-			s.NoError(err)
-
-			res, err = query.Query().Where("name = ?", "restore_user1").Restore(&User{})
-			s.Equal(int64(0), res.RowsAffected)
-			s.NoError(err)
-
-			res, err = query.Query().WithTrashed().Where("name = ?", "restore_user1").Restore(&User{})
-			s.Equal(int64(1), res.RowsAffected)
-			s.NoError(err)
-
-			res, err = query.Query().Model(&User{}).WithTrashed().Where("name = ?", "restore_user2").Restore()
-			s.Equal(int64(1), res.RowsAffected)
-			s.NoError(err)
-
-			res, err = query.Query().Model(users[2]).WithTrashed().Restore()
-			s.Equal(int64(1), res.RowsAffected)
-			s.NoError(err)
-
-			res, err = query.Query().WithTrashed().Restore(&users[3])
-			s.Equal(int64(1), res.RowsAffected)
-			s.NoError(err)
-
-			var count int64
-			s.NoError(query.Query().Model(&User{}).Where("avatar", "restore_avatar").Count(&count))
-			s.Equal(int64(4), count)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestSum() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "count_user", Avatar: "count_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "count_user", Avatar: "count_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var value float64
-			err := query.Query().Table("users").Sum("id", &value)
-			s.Nil(err)
-			s.True(value > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestToSql() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			switch driver {
-			case DriverPostgres:
-				s.Equal("SELECT * FROM \"users\" WHERE \"id\" = $1 AND \"users\".\"deleted_at\" IS NULL", query.Query().Where("id", 1).ToSql().Find(User{}))
-			case DriverSqlserver:
-				s.Equal("SELECT * FROM \"users\" WHERE \"id\" = @p1 AND \"users\".\"deleted_at\" IS NULL", query.Query().Where("id", 1).ToSql().Find(User{}))
-			default:
-				s.Equal("SELECT * FROM `users` WHERE `id` = ? AND `users`.`deleted_at` IS NULL", query.Query().Where("id", 1).ToSql().Find(User{}))
-			}
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestToRawSql() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			switch driver {
-			case DriverPostgres:
-				s.Equal("SELECT * FROM \"users\" WHERE \"id\" = 1 AND \"users\".\"deleted_at\" IS NULL", query.Query().Where("id", 1).ToRawSql().Find(User{}))
-			case DriverSqlserver:
-				s.Equal("SELECT * FROM \"users\" WHERE \"id\" = $1$ AND \"users\".\"deleted_at\" IS NULL", query.Query().Where("id", 1).ToRawSql().Find(User{}))
-			default:
-				s.Equal("SELECT * FROM `users` WHERE `id` = 1 AND `users`.`deleted_at` IS NULL", query.Query().Where("id", 1).ToRawSql().Find(User{}))
-			}
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestTransactionSuccess() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "transaction_success_user", Avatar: "transaction_success_avatar"}
-			user1 := User{Name: "transaction_success_user1", Avatar: "transaction_success_avatar1"}
-			tx, err := query.Query().Begin()
-			s.Nil(err)
-			s.Nil(tx.Create(&user))
-			s.Nil(tx.Create(&user1))
-			s.Nil(tx.Commit())
-
-			var user2, user3 User
-			s.Nil(query.Query().Find(&user2, user.ID))
-			s.Nil(query.Query().Find(&user3, user1.ID))
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestTransactionError() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "transaction_error_user", Avatar: "transaction_error_avatar"}
-			user1 := User{Name: "transaction_error_user1", Avatar: "transaction_error_avatar1"}
-			tx, err := query.Query().Begin()
-			s.Nil(err)
-			s.Nil(tx.Create(&user))
-			s.Nil(tx.Create(&user1))
-			s.Nil(tx.Rollback())
-
-			var users []User
-			s.Nil(query.Query().Where("name = ? or name = ?", "transaction_error_user", "transaction_error_user1").Find(&users))
-			s.Equal(0, len(users))
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestUpdate() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "update single column, success",
-				setup: func() {
-					users := []User{{Name: "updates_single_name", Avatar: "updates_single_avatar"}, {Name: "updates_single_name", Avatar: "updates_single_avatar1"}}
-					s.Nil(query.Query().Create(&users))
-					s.True(users[0].ID > 0)
-					s.True(users[1].ID > 0)
-
-					res, err := query.Query().Model(&User{}).Where("name = ?", "updates_single_name").Update("avatar", "update_single_avatar2")
-					s.Equal(int64(2), res.RowsAffected)
-					s.Nil(err)
-
-					var user2 User
-					s.Nil(query.Query().Find(&user2, users[0].ID))
-					s.Equal("update_single_avatar2", user2.Avatar)
-					var user3 User
-					s.Nil(query.Query().Find(&user3, users[1].ID))
-					s.Equal("update_single_avatar2", user3.Avatar)
-				},
-			},
-			{
-				name: "update columns by map, success",
-				setup: func() {
-					users := []User{{Name: "update_map_name", Avatar: "update_map_avatar"}, {Name: "update_map_name", Avatar: "update_map_avatar1"}}
-					s.Nil(query.Query().Create(&users))
-					s.True(users[0].ID > 0)
-					s.True(users[1].ID > 0)
-
-					res, err := query.Query().Model(&User{}).Where("name = ?", "update_map_name").Update(map[string]any{
-						"avatar": "update_map_avatar2",
-					})
-					s.Equal(int64(2), res.RowsAffected)
-					s.Nil(err)
-
-					var user2 User
-					s.Nil(query.Query().Find(&user2, users[0].ID))
-					s.Equal("update_map_avatar2", user2.Avatar)
-					var user3 User
-					s.Nil(query.Query().Find(&user3, users[0].ID))
-					s.Equal("update_map_avatar2", user3.Avatar)
-				},
-			},
-			{
-				name: "update columns by model, success",
-				setup: func() {
-					users := []User{{Name: "update_model_name", Avatar: "update_model_avatar"}, {Name: "update_model_name", Avatar: "update_model_avatar1"}}
-					s.Nil(query.Query().Create(&users))
-					s.True(users[0].ID > 0)
-					s.True(users[1].ID > 0)
-
-					res, err := query.Query().Model(&User{}).Where("name = ?", "update_model_name").Update(User{Avatar: "update_model_avatar2"})
-					s.Equal(int64(2), res.RowsAffected)
-					s.Nil(err)
-
-					var user2 User
-					s.Nil(query.Query().Find(&user2, users[0].ID))
-					s.Equal("update_model_avatar2", user2.Avatar)
-					var user3 User
-					s.Nil(query.Query().Find(&user3, users[0].ID))
-					s.Equal("update_model_avatar2", user3.Avatar)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
-	}
-}
-
-func (s *OrmTestSuite) TestUpdateOrCreate() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			var user User
-			err := query.Query().UpdateOrCreate(&user, User{Name: "update_or_create_user"}, User{Avatar: "update_or_create_avatar"})
-			s.Nil(err)
-			s.True(user.ID > 0)
-
-			var user1 User
-			err = query.Query().Where("name", "update_or_create_user").Find(&user1)
-			s.Nil(err)
-			s.True(user1.ID > 0)
-
-			var user2 User
-			err = query.Query().UpdateOrCreate(&user2, User{Name: "update_or_create_user"}, User{Avatar: "update_or_create_avatar1"})
-			s.Nil(err)
-			s.True(user2.ID > 0)
-			s.Equal("update_or_create_avatar1", user2.Avatar)
-
-			var user3 User
-			err = query.Query().Where("avatar", "update_or_create_avatar1").Find(&user3)
-			s.Nil(err)
-			s.True(user3.ID > 0)
-
-			var count int64
-			err = query.Query().Model(User{}).Where("name", "update_or_create_user").Count(&count)
-			s.Nil(err)
-			s.Equal(int64(1), count)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestWhere() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "where_user", Avatar: "where_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "where_user1", Avatar: "where_avatar1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var user2 []User
-			s.Nil(query.Query().Where("name = ?", "where_user").OrWhere("avatar = ?", "where_avatar1").Find(&user2))
-			s.Equal(2, len(user2))
-
-			var user3 User
-			s.Nil(query.Query().Where("name = 'where_user'").Find(&user3))
-			s.True(user3.ID > 0)
-
-			var user4 User
-			s.Nil(query.Query().Where("name", "where_user").Find(&user4))
-			s.True(user4.ID > 0)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestWhereIn() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "where_in_user", Avatar: "where_in_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "where_in_user_1", Avatar: "where_in_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().WhereIn("id", []any{user.ID, user1.ID}).Find(&users))
-			s.True(len(users) == 2)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestOrWhereIn() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "where_in_user", Avatar: "where_in_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "where_in_user_1", Avatar: "where_in_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().Where("id = ?", -1).OrWhereIn("id", []any{user.ID, user1.ID}).Find(&users))
-			s.True(len(users) == 2)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestWhereNotIn() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "where_in_user", Avatar: "where_in_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "where_in_user_1", Avatar: "where_in_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			user2 := User{Name: "where_in_user_2", Avatar: "where_in_avatar_2"}
-			s.Nil(query.Query().Create(&user2))
-			s.True(user2.ID > 0)
-
-			var user3 User
-			s.Nil(query.Query().Where("id = ?", user2.ID).WhereNotIn("id", []any{user.ID, user1.ID}).First(&user3))
-			s.True(user3.ID == user2.ID)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestOrWhereNotIn() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "where_in_user", Avatar: "where_in_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "where_in_user_1", Avatar: "where_in_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			user2 := User{Name: "where_in_user_2", Avatar: "where_in_avatar_2"}
-			s.Nil(query.Query().Create(&user2))
-			s.True(user2.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().Where("id = ?", -1).OrWhereNotIn("id", []any{user.ID, user1.ID}).Find(&users))
-			var user2Found bool
-			for _, user := range users {
-				if user.ID == user2.ID {
-					user2Found = true
-				}
-			}
-			s.True(user2Found)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestWhereBetween() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "where_between_user", Avatar: "where_between_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "where_between_user_1", Avatar: "where_between_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			user2 := User{Name: "where_between_user_2", Avatar: "where_between_avatar_2"}
-			s.Nil(query.Query().Create(&user2))
-			s.True(user2.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().WhereBetween("id", user.ID, user2.ID).Find(&users))
-			s.True(len(users) == 3)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestWhereNotBetween() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "where_not_between_user", Avatar: "where_not_between_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "where_not_between_user", Avatar: "where_not_between_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			user2 := User{Name: "where_not_between_user", Avatar: "where_not_between_avatar_2"}
-			s.Nil(query.Query().Create(&user2))
-			s.True(user2.ID > 0)
-
-			user3 := User{Name: "where_not_between_user", Avatar: "where_not_between_avatar_2"}
-			s.Nil(query.Query().Create(&user3))
-			s.True(user3.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().Where("name = ?", "where_not_between_user").WhereNotBetween("id", user.ID, user2.ID).Find(&users))
-			s.True(len(users) == 1)
-			s.True(users[0].ID == user3.ID)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestOrWhereBetween() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "or_where_between_user", Avatar: "or_where_between_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "or_where_between_user_1", Avatar: "or_where_between_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			user2 := User{Name: "or_where_between_user_2", Avatar: "or_where_between_avatar_2"}
-			s.Nil(query.Query().Create(&user2))
-			s.True(user2.ID > 0)
-
-			user3 := User{Name: "or_where_between_user_3", Avatar: "or_where_between_avatar_3"}
-			s.Nil(query.Query().Create(&user3))
-			s.True(user3.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().Where("name = ?", "or_where_between_user_3").OrWhereBetween("id", user.ID, user2.ID).Find(&users))
-			s.True(len(users) == 4)
-		})
+	for connection := range s.queries {
+		s.NotNil(s.orm.Connection(connection).Factory())
 	}
 }
-
-func (s *OrmTestSuite) TestOrWhereNotBetween() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "or_where_between_user", Avatar: "or_where_between_avatar"}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
-
-			user1 := User{Name: "or_where_between_user_1", Avatar: "or_where_between_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
-
-			user2 := User{Name: "or_where_between_user_2", Avatar: "or_where_between_avatar_2"}
-			s.Nil(query.Query().Create(&user2))
-			s.True(user2.ID > 0)
-
-			user3 := User{Name: "or_where_between_user_3", Avatar: "or_where_between_avatar_3"}
-			s.Nil(query.Query().Create(&user3))
-			s.True(user3.ID > 0)
-
-			var users []User
-			s.Nil(query.Query().Where("name = ?", "or_where_between_user_3").OrWhereNotBetween("id", user.ID, user2.ID).Find(&users))
-			s.True(len(users) >= 1)
-		})
-	}
-}
-
-func (s *OrmTestSuite) TestWhereNull() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			bio := "where_null_bio"
-			user := User{Name: "where_null_user", Avatar: "where_null_avatar", Bio: &bio}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
 
-			user1 := User{Name: "where_null_user", Avatar: "where_null_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
+func (s *OrmSuite) TestObserve() {
+	s.orm.Observe(Role{}, &UserObserver{})
 
-			var users []User
-			s.Nil(query.Query().Where("name = ?", "where_null_user").WhereNull("bio").Find(&users))
-			s.True(len(users) == 1)
-			s.True(users[0].ID == user1.ID)
-		})
+	for connection := range s.queries {
+		role := Role{Name: "observer_name"}
+		s.EqualError(s.orm.Connection(connection).Query().Create(&role), "error")
 	}
 }
 
-func (s *OrmTestSuite) TestOrWhereNull() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			bio := "or_where_null_bio"
-			user := User{Name: "or_where_null_user", Avatar: "or_where_null_avatar", Bio: &bio}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
+func (s *OrmSuite) TestTransactionSuccess() {
+	for connection := range s.queries {
+		role := Role{Name: "transaction_success_role", Avatar: "transaction_success_avatar"}
+		role1 := Role{Name: "transaction_success_role1", Avatar: "transaction_success_avatar1"}
+		s.Nil(s.orm.Connection(connection).Transaction(func(tx contractsorm.Query) error {
+			s.Nil(tx.Create(&role))
+			s.Nil(tx.Create(&role1))
 
-			user1 := User{Name: "or_where_null_user_1", Avatar: "or_where_null_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
+			return nil
+		}))
 
-			var users []User
-			s.Nil(query.Query().Where("name = ?", "or_where_null_user").OrWhereNull("bio").Find(&users))
-			s.True(len(users) >= 2)
-		})
+		var role2, role3 Role
+		s.Nil(s.orm.Connection(connection).Query().Find(&role2, role.ID))
+		s.Nil(s.orm.Connection(connection).Query().Find(&role3, role1.ID))
 	}
 }
 
-func (s *OrmTestSuite) TestWhereNotNull() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			bio := "where_not_null_bio"
-			user := User{Name: "where_not_null_user", Avatar: "where_not_null_avatar", Bio: &bio}
-			s.Nil(query.Query().Create(&user))
-			s.True(user.ID > 0)
+func (s *OrmSuite) TestTransactionError() {
+	for connection := range s.queries {
+		s.NotNil(s.orm.Connection(connection).Transaction(func(tx contractsorm.Query) error {
+			role := Role{Name: "transaction_error_role", Avatar: "transaction_error_avatar"}
+			s.Nil(tx.Create(&role))
 
-			user1 := User{Name: "where_not_null_user", Avatar: "where_not_null_avatar_1"}
-			s.Nil(query.Query().Create(&user1))
-			s.True(user1.ID > 0)
+			role1 := Role{Name: "transaction_error_role1", Avatar: "transaction_error_avatar1"}
+			s.Nil(tx.Create(&role1))
 
-			var users []User
-			s.Nil(query.Query().Where("name = ?", "where_not_null_user").WhereNotNull("bio").Find(&users))
-			s.True(len(users) == 1)
-			s.True(users[0].ID == user.ID)
-		})
-	}
-}
+			return errors.New("error")
+		}))
 
-func (s *OrmTestSuite) TestWithoutEvents() {
-	for _, query := range s.queries {
-		tests := []struct {
-			name  string
-			setup func()
-		}{
-			{
-				name: "success",
-				setup: func() {
-					user := User{Name: "event_save_without_name", Avatar: "without_events_avatar"}
-					s.Nil(query.Query().WithoutEvents().Save(&user))
-					s.True(user.ID > 0)
-					s.Equal("without_events_avatar", user.Avatar)
-
-					var user1 User
-					s.Nil(query.Query().Find(&user1, user.ID))
-					s.Equal("event_save_without_name", user1.Name)
-					s.Equal("without_events_avatar", user1.Avatar)
-				},
-			},
-		}
-		for _, test := range tests {
-			s.Run(test.name, func() {
-				test.setup()
-			})
-		}
+		var roles []Role
+		s.Nil(s.orm.Connection(connection).Query().Find(&roles))
+		s.Equal(0, len(roles))
 	}
 }
-
-func (s *OrmTestSuite) TestWith() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "with_user", Address: &Address{
-				Name: "with_address",
-			}, Books: []*Book{{
-				Name: "with_book0",
-			}, {
-				Name: "with_book1",
-			}}}
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-			s.True(user.Address.ID > 0)
-			s.True(user.Books[0].ID > 0)
-			s.True(user.Books[1].ID > 0)
 
-			tests := []struct {
-				description string
-				setup       func(description string)
-			}{
-				{
-					description: "simple",
-					setup: func(description string) {
-						var user1 User
-						s.Nil(query.Query().With("Address").With("Books").Find(&user1, user.ID))
-						s.True(user1.ID > 0)
-						s.True(user1.Address.ID > 0)
-						s.True(user1.Books[0].ID > 0)
-						s.True(user1.Books[1].ID > 0)
-					},
-				},
-				{
-					description: "with simple conditions",
-					setup: func(description string) {
-						var user1 User
-						s.Nil(query.Query().With("Books", "name = ?", "with_book0").Find(&user1, user.ID))
-						s.True(user1.ID > 0)
-						s.Nil(user1.Address)
-						s.Equal(1, len(user1.Books))
-						s.Equal("with_book0", user1.Books[0].Name)
-					},
-				},
-				{
-					description: "with func conditions",
-					setup: func(description string) {
-						var user1 User
-						s.Nil(query.Query().With("Books", func(query contractsorm.Query) contractsorm.Query {
-							return query.Where("name = ?", "with_book0").Select("id", "user_id", "name")
-						}).Find(&user1, user.ID))
-						s.True(user1.ID > 0)
-						s.Nil(user1.Address)
-						s.Equal(1, len(user1.Books))
-						s.Equal("with_book0", user1.Books[0].Name)
-					},
-				},
-			}
-			for _, test := range tests {
-				test.setup(test.description)
-			}
-		})
-	}
-}
+func (s *OrmSuite) TestWithContext() {
+	s.orm.Observe(Role{}, &UserObserver{})
+	ctx := context.WithValue(context.Background(), testContextKey, "with_context_goravel")
+	role := Role{Name: "with_context_name"}
 
-func (s *OrmTestSuite) TestWithNesting() {
-	for driver, query := range s.queries {
-		s.Run(driver, func() {
-			user := User{Name: "with_nesting_user", Books: []*Book{{
-				Name:   "with_nesting_book0",
-				Author: &Author{Name: "with_nesting_author0"},
-			}, {
-				Name:   "with_nesting_book1",
-				Author: &Author{Name: "with_nesting_author1"},
-			}}}
-			s.Nil(query.Query().Select(gorm.Associations).Create(&user))
-			s.True(user.ID > 0)
-			s.True(user.Books[0].ID > 0)
-			s.True(user.Books[0].Author.ID > 0)
-			s.True(user.Books[1].ID > 0)
-			s.True(user.Books[1].Author.ID > 0)
+	// Call Query directly
+	err := s.orm.WithContext(ctx).Query().Create(&role)
+	s.Nil(err)
+	s.Equal("with_context_name", role.Name)
+	s.Equal("with_context_goravel", role.Avatar)
 
-			var user1 User
-			s.Nil(query.Query().With("Books.Author").Find(&user1, user.ID))
-			s.True(user1.ID > 0)
-			s.Equal("with_nesting_user", user1.Name)
-			s.True(user1.Books[0].ID > 0)
-			s.Equal("with_nesting_book0", user1.Books[0].Name)
-			s.True(user1.Books[0].Author.ID > 0)
-			s.Equal("with_nesting_author0", user1.Books[0].Author.Name)
-			s.True(user1.Books[1].ID > 0)
-			s.Equal("with_nesting_book1", user1.Books[1].Name)
-			s.True(user1.Books[1].Author.ID > 0)
-			s.Equal("with_nesting_author1", user1.Books[1].Author.Name)
-		})
+	// Call Connection, then call WithContext
+	for connection := range s.queries {
+		role.ID = 0
+		role.Avatar = ""
+		err := s.orm.Connection(connection).WithContext(ctx).Query().Create(&role)
+		s.Nil(err)
+		s.Equal("with_context_name", role.Name)
+		s.Equal("with_context_goravel", role.Avatar)
 	}
-}
-
-// func TestCustomConnection(t *testing.T) {
-// 	if env.IsWindows() {
-// 		t.Skip("Skip test that using Docker")
-// 	}
-
-// 	postgresDocker := supportdocker.Postgres()
-// 	require.NoError(t, postgresDocker.Ready())
-
-// 	postgresQuery := NewTestQuery(postgresDocker)
-// 	postgresQuery.CreateTable(TestTableReviews, TestTableProducts)
-
-// 	sqliteDocker := supportdocker.Sqlite()
-// 	sqliteQuery := NewTestQuery(sqliteDocker)
-// 	sqliteQuery.CreateTable(TestTableReviews, TestTableProducts)
-
-// 	query := postgresQuery.Query()
-
-// 	review := Review{Body: "create_review"}
-// 	assert.Nil(t, query.Create(&review))
-// 	assert.True(t, review.ID > 0)
-
-// 	var review1 Review
-// 	assert.Nil(t, query.Where("body", "create_review").First(&review1))
-// 	assert.True(t, review1.ID > 0)
-
-// 	mockCommonConnection(postgresQuery.MockConfig(), sqliteQuery, "sqlite")
-
-// 	product := Product{Name: "create_product"}
-// 	assert.Nil(t, query.Create(&product))
-// 	assert.True(t, product.ID > 0)
 
-// 	var product1 Product
-// 	assert.Nil(t, query.Where("name", "create_product").First(&product1))
-// 	assert.True(t, product1.ID > 0)
-
-// 	var product2 Product
-// 	assert.Nil(t, query.Where("name", "create_product1").First(&product2))
-// 	assert.True(t, product2.ID == 0)
-
-// 	mockCommonConnection(postgresQuery.MockConfig(), postgresQuery, "dummy")
-
-// 	person := Person{Name: "create_person"}
-// 	assert.NotNil(t, query.Create(&person))
-// 	assert.True(t, person.ID == 0)
-
-// 	assert.NoError(t, sqliteDocker.Shutdown())
-// }
-
-// func TestReadWriteSeparate(t *testing.T) {
-// 	if env.IsWindows() {
-// 		t.Skip("Skip test that using Docker")
-// 	}
-
-// 	dbs := NewTestQueries().QueriesOfReadWrite()
-
-// 	for drive, db := range dbs {
-// 		t.Run(drive.String(), func(t *testing.T) {
-// 			var (
-// 				mixQuery contractsorm.Query
-// 				err      error
-// 			)
-// 			if drive == database.DriverSqlite {
-// 				mixQuery, err = db["write"].QueryOfReadWrite(db["read"].Docker().Config())
-// 			} else {
-// 				mixQuery, err = db["write"].QueryOfReadWrite(db["read"].Docker().Config())
-// 			}
-
-// 			require.NoError(t, err)
-
-// 			db["read"].CreateTable(TestTableUsers)
-// 			db["write"].CreateTable(TestTableUsers)
-
-// 			user := User{Name: "user"}
-// 			assert.Nil(t, mixQuery.Create(&user))
-// 			assert.True(t, user.ID > 0)
-
-// 			var user2 User
-// 			assert.Nil(t, mixQuery.Find(&user2, user.ID))
-// 			assert.True(t, user2.ID == 0)
-
-// 			var user3 User
-// 			assert.Nil(t, db["read"].Query().Find(&user3, user.ID))
-// 			assert.True(t, user3.ID == 0)
-
-// 			var user4 User
-// 			assert.Nil(t, db["write"].Query().Find(&user4, user.ID))
-// 			assert.True(t, user4.ID > 0)
-// 		})
-// 	}
-
-// 	assert.NoError(t, dbs[database.DriverSqlite]["read"].Docker().Shutdown())
-// 	assert.NoError(t, dbs[database.DriverSqlite]["write"].Docker().Shutdown())
-// }
-
-// func TestTablePrefixAndSingular(t *testing.T) {
-// 	if env.IsWindows() {
-// 		t.Skip("Skip test that using Docker")
-// 	}
-
-// 	dbs := NewTestQueries().QueriesWithPrefixAndSingular()
-
-// 	for drive, db := range dbs {
-// 		t.Run(drive.String(), func(t *testing.T) {
-// 			db.CreateTable(TestTableGoravelUser)
-
-// 			user := User{Name: "user"}
-// 			assert.Nil(t, db.Query().Create(&user))
-// 			assert.True(t, user.ID > 0)
-
-// 			var user1 User
-// 			assert.Nil(t, db.Query().Find(&user1, user.ID))
-// 			assert.True(t, user1.ID > 0)
-// 		})
-// 	}
-
-// 	if dbs[database.DriverSqlite] != nil {
-// 		assert.NoError(t, dbs[database.DriverSqlite].Docker().Shutdown())
-// 	}
-// }
-
-// func TestPostgresSchema(t *testing.T) {
-// 	if env.IsWindows() {
-// 		t.Skip("Skip test that using Docker")
-// 	}
-
-// 	postgresDocker := supportdocker.Postgres()
-// 	require.NoError(t, postgresDocker.Ready())
-
-// 	testQuery := NewTestQueryWithSchema(postgresDocker, "goravel")
-// 	testQuery.CreateTable(TestTableUsers)
-
-// 	user := User{Name: "first_user"}
-// 	assert.Nil(t, testQuery.Query().Create(&user))
-// 	assert.True(t, user.ID > 0)
-
-// 	var user1 User
-// 	assert.Nil(t, testQuery.Query().Where("name", "first_user").First(&user1))
-// 	assert.True(t, user1.ID > 0)
-// }
-
-// func TestSqlserverSchema(t *testing.T) {
-// 	if env.IsWindows() {
-// 		t.Skip("Skip test that using Docker")
-// 	}
-
-// 	sqlserverDocker := supportdocker.Sqlserver()
-// 	require.NoError(t, sqlserverDocker.Ready())
-
-// 	testQuery := NewTestQueryWithSchema(sqlserverDocker, "goravel")
-// 	testQuery.CreateTable(TestTableSchema)
-
-// 	schema := Schema{Name: "first_schema"}
-// 	assert.Nil(t, testQuery.Query().Create(&schema))
-// 	assert.True(t, schema.ID > 0)
-
-// 	var schema1 Schema
-// 	assert.Nil(t, testQuery.Query().Where("name", "first_schema").First(&schema1))
-// 	assert.True(t, schema1.ID > 0)
-// }
-
-func paginator(page string, limit string) func(methods contractsorm.Query) contractsorm.Query {
-	return func(query contractsorm.Query) contractsorm.Query {
-		page, _ := strconv.Atoi(page)
-		limit, _ := strconv.Atoi(limit)
-		offset := (page - 1) * limit
-
-		return query.Offset(offset).Limit(limit)
+	// Call WithContext, then call Connection
+	for connection := range s.queries {
+		role.ID = 0
+		role.Avatar = ""
+		err := s.orm.WithContext(ctx).Connection(connection).Query().Create(&role)
+		s.Nil(err)
+		s.Equal("with_context_name", role.Name)
+		s.Equal("with_context_goravel", role.Avatar)
 	}
 }
-
-// func mockCommonConnection(mockConfig *mocksconfig.Config, testQuery *TestQuery, connection string) {
-// 	mockDriver := getMockDriver(testQuery.Docker(), mockConfig, connection)
-// 	mockDriver.Common()
-// }
